### 2026-03-29 #10316 @editorialbot
[PRE REVIEW]: LinearMPC.jl: A Julia Package for Embedded Linear Model Predictive Control
**Submitting author:** @darnstrom (<a href="http://orcid.org/0000-0003-0970-0620">Daniel Arnström</a>)
**Repository:** https://github.com/darnstrom/LinearMPC.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.9.0
