### 2026-04-01 #10316 @danielskatz
@jolars - you can use the command `@editorialbot assign me as editor` to assign yourself
