### 2026-04-01 #10316 @jolars
:wave: @jbcaillau, @imciner2, @baggepinnen, @franckgaga, @alexdada555 would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
