### 2026-04-01 #10316 @franckgaga
I can review but:

1. Am I biased or in conflict of interest if I am indirectly involved in the development of this package? I developed an interface to ModelPredictiveControl.jl, submitted a PR and helped in the debugging.
2. If no, I'm leaving to Mexico tomorrow and will I start a new job when I will be back in Canada, so I won't have time for this in the next two weeks or so.
