### 2026-04-05 eee47c9
Add tabs for Python code in examples
