### 2026-04-05 4807d65
[DOCS] Add tabs for Python code in examples (#92)
* Add tabs for Python code in examples

* Cleanup in docs
