+++ docs/make.jl
@@ -2,8 +2,153 @@ push!(LOAD_PATH,"../src/")
 using LinearMPC
 using Documenter
 
+# ============================================================
+# Tab block preprocessor
+#
+# Introduces three custom fenced-code block types so that Julia
+# code only needs to be written *once* per example:
+#
+#   ```@tab
+#   # julia
+#   <julia code shown in the Julia tab>
+#   # python
+#   <python code shown in the Python tab>
+#   ```
+#   (display-only: no code is executed)
+#
+#   ```@tabexample name
+#   # julia
+#   <julia code shown in the Julia tab AND executed via @example>
+#   # exec-only          (optional)
+#   <extra lines executed but not shown in the tab, e.g. ylims!>
+#   # python
+#   <python code shown in the Python tab>
+#   ```
+#
+#   ```@tabsetup name
+#   # julia
+#   <julia code shown in the Julia tab AND executed via @setup>
+#   # python
+#   <python code shown in the Python tab>
+#   ```
+#
+# The preprocessor expands each block into:
+#   • a ```@raw html``` lang-switcher showing the code, and
+#   • (for @tabexample/@tabsetup) a ```@example name``` /
+#     ```@setup name``` block that runs the Julia code
+#     (all lines auto-hidden so only output shows).
+#
+# Files are modified in-place before makedocs and restored in the
+# finally block below.
+# ============================================================
+
+function _escape_html(s::AbstractString)::String
+    s = replace(s, "&" => "&amp;")
+    s = replace(s, "<" => "&lt;")
+    s = replace(s, ">" => "&gt;")
+    return String(rstrip(s))
+end
+
+function _make_lang_switcher_html(julia_code::AbstractString, python_code::AbstractString)::String
+    j = _escape_html(julia_code)
+    p = _escape_html(python_code)
+    return """<div class="lang-switcher">
+<div class="lang-switcher-tabs">
+<button class="lang-switcher-tab active" data-lang="julia"><img src="../../assets/julia.svg" alt="" class="lang-icon"> Julia</button>
+<button class="lang-switcher-tab" data-lang="python"><img src="../../assets/python.svg" alt="" class="lang-icon"> Python</button>
+</div>
+<div class="lang-switcher-content active" data-lang="julia"><pre><code class="language-julia">$(j)</code></pre></div>
+<div class="lang-switcher-content" data-lang="python"><pre><code class="language-python">$(p)</code></pre></div>
+</div>"""
+end
+
+# Append " # hide" to every non-blank line that doesn't already have it.
+function _add_hide(code::AbstractString)::String
+    return join(map(split(code, "\n")) do line
+        stripped = rstrip(line)
+        isempty(stripped) || endswith(stripped, "# hide") ? line : line * " # hide"
+    end, "\n")
+end
+
+# Parse a tab block body into (julia_display, julia_exec_only, python) sections.
+function _parse_sections(body::AbstractString)
+    julia_lines, exec_lines, python_lines = String[], String[], String[]
+    section = :none
+    for line in split(body, "\n")
+        s = strip(line)
+        if     s == "# julia";     section = :julia
+        elseif s == "# exec-only"; section = :exec_only
+        elseif s == "# python";    section = :python
+        elseif section == :julia;      push!(julia_lines, line)
+        elseif section == :exec_only;  push!(exec_lines, line)
+        elseif section == :python;     push!(python_lines, line)
+        end
+    end
+    return rstrip(join(julia_lines, "\n")),
+           rstrip(join(exec_lines, "\n")),
+           rstrip(join(python_lines, "\n"))
+end
+
+function _expand_tab_blocks(content::String)::String
[…]
