+++ docs/src/manual/linear_cost.md
@@ -10,40 +10,54 @@ This example demonstrates using linear cost to optimize heating while accounting
 
 First, we define a function to generate realistic electricity prices with daily patterns:
 
-```@example linear_cost
+```@tabsetup linear_cost
+# julia
 function generate_test_prices(n_days=7)
     n_per_day = 24
     n = n_days * n_per_day
-
     prices = zeros(n)
-
     for day in 1:n_days
         t = range(0, 24, length=n_per_day)
-
         daily_factor = 0.8 + 0.4 * rand()
-
         base = 0.5 .+ 0.2 * sin.(2π * (t .- 6) / 24)
-
         morning_shift = 8 + randn() * 0.5
         evening_shift = 18 + randn() * 0.5
         morning_peak = (0.4 + 0.2 * rand()) * exp.(-((t .- morning_shift) .^ 2) / 2)
         evening_peak = (0.3 + 0.2 * rand()) * exp.(-((t .- evening_shift) .^ 2) / 3)
         night_dip = -0.1 * exp.(-((t .- 3) .^ 2) / 4)
-
         day_prices = daily_factor * (base + morning_peak + evening_peak + night_dip)
         day_prices = day_prices .+ 0.05 * randn(n_per_day)
-
         idx_start = (day - 1) * n_per_day + 1
         idx_end = day * n_per_day
         prices[idx_start:idx_end] = day_prices
     end
-
     prices = max.(prices, 0.1)
     prices = 0.2 .+ 1.5 * (prices .- minimum(prices)) / (maximum(prices) - minimum(prices))
-
     return prices
 end
-nothing # hide
+# python
+import numpy as np
+
+def generate_test_prices(n_days=7):
+    n_per_day = 24
+    n = n_days * n_per_day
+    prices = np.zeros(n)
+    for day in range(n_days):
+        t = np.linspace(0, 24, n_per_day)
+        daily_factor = 0.8 + 0.4 * np.random.rand()
+        base = 0.5 + 0.2 * np.sin(2 * np.pi * (t - 6) / 24)
+        morning_shift = 8 + np.random.randn() * 0.5
+        evening_shift = 18 + np.random.randn() * 0.5
+        morning_peak = (0.4 + 0.2 * np.random.rand()) * np.exp(-((t - morning_shift) ** 2) / 2)
+        evening_peak = (0.3 + 0.2 * np.random.rand()) * np.exp(-((t - evening_shift) ** 2) / 3)
+        night_dip = -0.1 * np.exp(-((t - 3) ** 2) / 4)
+        day_prices = daily_factor * (base + morning_peak + evening_peak + night_dip)
+        day_prices = day_prices + 0.05 * np.random.randn(n_per_day)
+        idx_start = day * n_per_day
+        prices[idx_start:idx_start + n_per_day] = day_prices
+    prices = np.maximum(prices, 0.1)
+    prices = 0.2 + 1.5 * (prices - prices.min()) / (prices.max() - prices.min())
+    return prices
 ```
 
 ### Room Temperature Model
@@ -54,7 +68,8 @@ T_{k+1} = a \cdot T_k + b \cdot u_k
 ```
 where $T$ is room temperature, $u$ is heating power, $a < 1$ represents heat loss, and $b$ is the heating efficiency.
 
-```@example linear_cost
+```@tabsetup linear_cost
+# julia
 using LinearMPC
 using Random
 Random.seed!(0)
@@ -70,20 +85,42 @@ C = [1.0;;]
 Np = 24  # 24-hour prediction horizon
 
 mpc = LinearMPC.MPC(A, B; C, Np=Np, Nc=Np)
[…]
