### 2026-04-08 #10384 @goerz
@darnstrom Quick question about your affiliation ("Independent Researcher"), as it appears on the paper: [Your website](https://darnstrom.github.io) has you as a postdoc at Uppsala University, and the package certainly seems "academic" in nature. Shouldn't you use that as your affiliation? Even if you've recently left the position, if any development of this package was part of your postdoc work, it seems more appropriate to still use the affiliation. You may want to check with your PI on this – and, in fact, you may want to include your PI as a second author. People, fields, and journals have various stances on the "Directed the project" authorship contribution of a PI, but at least in my own field of physics, the PI is always the last author on a paper, and it would raise eyebrows (and cause problems for them) if they were omitted. Also for you, having an affiliation of the paper probably gives it more "respectability". Probably not as much for a software paper as it would for a research paper, where "no affiliation" would be a red flag, but still.

It's up to you, of course
