### 2026-04-08 #10384 @darnstrom
> [@darnstrom](https://github.com/darnstrom) Quick question about your affiliation ("Independent Researcher"), as it appears on the paper: [Your website](https://darnstrom.github.io) has you as a postdoc at Uppsala University, and the package certainly seems "academic" in nature. Shouldn't you use that as your affiliation? Even if you've recently left the position, if any development of this package was part of your postdoc work, it seems more appropriate to still use the affiliation. You may want to check with your PI on this – and, in fact, you may want to include your PI as a second author. People, fields, and journals have various stances on the "Directed the project" authorship contribution of a PI, but at least in my own field of physics, the PI is always the last author on a paper, and it would raise eyebrows (and cause problems for them) if they were omitted. Also for you, having an affiliation of the paper probably gives it more "respectability". Probably not as much for a software paper as it would for a research paper, where "no affiliation" would be a red flag, but still.

That the affiliation "Independent Researcher" can be red flag is a valid point @goerz. Making sure that the right affiliation is disclosed is important, so thanks for bringing it up!
I classify LinearMPC.jl as a "side project" and it is not directly connected to the work I did in my previous postdoc position at UU. It is more connected to my PhD thesis rather than the research I did during my postdoc (still this package was not developed while I was a PhD student.) The main reason for putting "Independent researcher" is that all of the work on the package has been carried out by myself outside of my academic positions (although, the package is clearly related to my research, since it is an effort to make the results of my research more accessible.) Since I am currently not affiliated with any university, and that my current position is not related to this package, I felt that "No affiliation" was the most suitable in this instance.

Btw, thanks for reminding me that I need to update my personal website!
