### 2026-04-08 5eea08a
Add codecov CI integration and README status badges (#94)
* add codecov CI integration and README badges for CI and coverage

Agent-Logs-Url: https://github.com/darnstrom/LinearMPC.jl/sessions/eff6b934-afb4-4c74-92de-716f256a99b2

Co-authored-by: darnstrom <55484604+darnstrom@users.noreply.github.com>

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-authored-by: darnstrom <55484604+darnstrom@users.noreply.github.com>
