### 2026-04-08 fea8333
Add coverage-driven helper and recipe tests
Agent-Logs-Url: https://github.com/darnstrom/LinearMPC.jl/sessions/69379c13-6c7a-47d1-a32c-832e3fe1f7e1

Co-authored-by: darnstrom <55484604+darnstrom@users.noreply.github.com>
