+++ test/runtests.jl
@@ -3,8 +3,10 @@ using Test
 using LinearAlgebra
 using Statistics
 using Random
+using RecipesBase
 global templib
 Random.seed!(1234)
+RecipesBase.is_key_supported(::Symbol) = true
 
 @testset verbose = true "LinearMPC.jl" begin
     @testset "Basic setup" begin
@@ -870,5 +872,271 @@ Random.seed!(1234)
 
         @test sim_exp.ys[1,end] ≈ 10.0 atol=1e-4
         @test sim_exp.ys[2,end] ≈ 0.0 atol=1e-4
+    end
+
+    @testset "Model constructors and linearization helpers" begin
+        F = [1.0 0.2; 0.0 1.0]
+        G = [0.0; 1.0]
+        C = [1.0 0.0]
+
+        model = LinearMPC.Model(F, G; C, Gd=[0.5; -0.5])
+        @test size(model.Gd) == (2, 1)
+        @test size(model.Dd) == (1, 1)
+        @test model.true_dynamics([1.0, 2.0], [3.0], [4.0]) ≈ F * [1.0, 2.0] + G * [3.0] + model.Gd * [4.0]
+        @test model.true_h([1.0, 2.0], [3.0], [4.0]) ≈ C * [1.0, 2.0]
+
+        cmodel = LinearMPC.Model([0.0 1.0; 0.0 0.0], [0.0; 1.0], 0.1; C)
+        @test cmodel.F ≈ [1.0 0.1; 0.0 1.0] atol=1e-8
+        @test cmodel.G ≈ [0.005; 0.1] atol=1e-8
+
+        f(x, u, d) = [x[1] + 2x[2] + u[1] + 3d[1]; x[2] + 2u[1] - d[1]]
+        h(x, u, d) = [x[1] - d[1]]
+        x0 = [1.0, -1.0]
+        u0 = [0.5]
+        d0 = [2.0]
+        A, B, Bd, Cn, D, Dd, foff, hoff = LinearMPC.linearize(f, h, x0, u0; d=d0)
+        @test A ≈ [1.0 2.0; 0.0 1.0]
+        @test B ≈ [1.0; 2.0;;]
+        @test Bd ≈ [3.0; -1.0;;]
+        @test Cn ≈ [1.0 0.0]
+        @test D ≈ zeros(1, 1)
+        @test Dd ≈ [-1.0;;]
+        @test foff ≈ zeros(2)
+        @test hoff ≈ zeros(1)
+
+        nmodel = LinearMPC.Model(f, h, x0, u0; d=d0)
+        @test nmodel.F ≈ A
+        @test nmodel.G ≈ B
+        @test nmodel.Gd ≈ Bd
+        @test nmodel.C ≈ Cn
+
+        @test_throws ArgumentError LinearMPC.Model(F, G; C=ones(3, 3))
+        @test_throws ArgumentError LinearMPC.Model(f, (x, u, d) -> [x[1] + u[1]], x0, u0; d=d0)
+    end
+
+    @testset "Reference and linear-cost formatting helpers" begin
+        mpc = LinearMPC.MPC([1.0 1.0; 0.0 1.0], [0.0; 1.0]; C=[1.0 0.0; 0.0 1.0], Np=4, Nc=4)
+        set_objective!(mpc; Q=[1.0, 1.0], R=[1.0])
+
+        mpc.settings.reference_preview = true
+        @test LinearMPC.format_reference(mpc, [1.0, 2.0]) ≈ repeat([1.0, 2.0], mpc.Np)
[…]
