### 2026-04-08 9af6974
Scope recipe test shim to recipe testsets
Agent-Logs-Url: https://github.com/darnstrom/LinearMPC.jl/sessions/69379c13-6c7a-47d1-a32c-832e3fe1f7e1

Co-authored-by: darnstrom <55484604+darnstrom@users.noreply.github.com>
