+++ test/runtests.jl
@@ -955,71 +955,75 @@ Random.seed!(1234)
         @test_throws ArgumentError LinearMPC.format_linear_cost(varying, [1.0 2.0 3.0 4.0; 4.0 3.0 2.0 1.0])
     end
 
-    # RecipesBase expects plotting backends to register supported keys; these tests exercise recipes directly.
-    RecipesBase.is_key_supported(::Symbol) = true
+    @testset "Recipe helper coverage" begin
+        shim = @eval RecipesBase is_key_supported(::Symbol) = true
+        try
+            @testset "Simulation preview and recipe helpers" begin
+                rs = [1.0 2.0 3.0; 10.0 20.0 30.0]
+                @test LinearMPC.get_preview(rs, 1, 4) == [2.0 3.0 3.0 3.0; 20.0 30.0 30.0 30.0]
+                @test LinearMPC.get_reference_preview(rs, 1, 4) == LinearMPC.get_preview(rs, 1, 4)
+                ls = [1.0 2.0 3.0 4.0]
+                @test LinearMPC.get_linear_cost_preview(ls, 2, 3) == [2.0 3.0 4.0]
 
-    @testset "Simulation preview and recipe helpers" begin
-        rs = [1.0 2.0 3.0; 10.0 20.0 30.0]
-        @test LinearMPC.get_preview(rs, 1, 4) == [2.0 3.0 3.0 3.0; 20.0 30.0 30.0 30.0]
-        @test LinearMPC.get_reference_preview(rs, 1, 4) == LinearMPC.get_preview(rs, 1, 4)
-        ls = [1.0 2.0 3.0 4.0]
-        @test LinearMPC.get_linear_cost_preview(ls, 2, 3) == [2.0 3.0 4.0]
+                mpc = LinearMPC.MPC([1.0;;], [1.0;;]; C=[1.0;;], Np=3)
+                set_objective!(mpc; Q=[1.0], R=[1.0])
+                set_bounds!(mpc; umin=[-1.0], umax=[1.0])
+                setup!(mpc)
+                sim = LinearMPC.Simulation(mpc; x0=[0.0], r=[0.0], N=3)
+                attrs = RecipesBase.KW(:xids => [1])
+                series = RecipesBase.apply_recipe(attrs, sim)
+                @test length(series) == 6
+                @test attrs[:layout] == reshape([(1, 1), (1, 1), (1, 1)], 1, :)
+                @test series[1].args == (sim.ts, sim.rs[1, :])
+                @test series[end].args == (sim.ts, sim.xs[1, :])
+            end
 
-        mpc = LinearMPC.MPC([1.0;;], [1.0;;]; C=[1.0;;], Np=3)
-        set_objective!(mpc; Q=[1.0], R=[1.0])
-        set_bounds!(mpc; umin=[-1.0], umax=[1.0])
-        setup!(mpc)
-        sim = LinearMPC.Simulation(mpc; x0=[0.0], r=[0.0], N=3)
-        attrs = RecipesBase.KW(:xids => [1])
-        series = RecipesBase.apply_recipe(attrs, sim)
-        @test length(series) == 6
-        @test attrs[:layout] == reshape([(1, 1), (1, 1), (1, 1)], 1, :)
-        @test series[1].args == (sim.ts, sim.rs[1, :])
-        @test series[end].args == (sim.ts, sim.xs[1, :])
-    end
+            @testset "Explicit MPC parameter labeling and plotting helpers" begin
+                mpc = LinearMPC.MPC([1.0 1.0; 0.0 1.0], [0.0; 1.0]; Gd=[1.0; 0.0], C=[1.0 0.0; 0.0 1.0], Dd=[0.0; 1.0], Np=3)
+                set_objective!(mpc; Q=[1.0, 1.0], R=[1.0], Rr=[0.5])
[…]
