### 2026-04-08 dc9177b
Add installation instructions (#97)
