+++ README.md
@@ -20,6 +20,17 @@ $$
 where $\hat{x}$ is the current state and $r$ is the desired reference value of $Cx$.
 
 
+## Installation
+
+[As usual for a registered Julia package](https://docs.julialang.org/en/v1/stdlib/Pkg/), `LinearMPC` can be installed by typing
+
+```
+] add LinearMPC
+```
+
+in the Julia REPL.
+
+
 ## Example 
 The following code show a simple MPC example of controlling an inverted pendulum on a cart, inspired by [this](https://se.mathworks.com/help/mpc/ug/mpc-control-of-an-inverted-pendulum-on-a-cart.html) example in the Model Predictive Toolbox in MATLAB.
 ```julia
