### 2026-04-08 ae82dde
Add initial benchmarks in docs
