### 2026-04-08 b7dfe96
Add initial benchmarks in docs (#98)
