+++ src/observer.jl
@@ -10,6 +10,45 @@ struct KalmanFilter
     x::Vector{Float64} 
 end
 
+struct OffsetFreeObserver
+    estimator::KalmanFilter
+    C::Matrix{Float64}
+    Dd::Matrix{Float64}
+    h_offset::Vector{Float64}
+    nx::Int
+    nd_measured::Int
+    nd_offsetfree::Int
+    formulation::Symbol
+end
+
+function Base.propertynames(observer::OffsetFreeObserver, private::Bool=false)
+    public_names = (:estimator, :C, :Dd, :h_offset, :nx, :nd_measured,
+                    :nd_offsetfree, :formulation, :x, :d, :K)
+    return private ? (public_names..., fieldnames(KalmanFilter)...) : public_names
+end
+
+function Base.getproperty(observer::OffsetFreeObserver, name::Symbol)
+    if name === :x
+        nx = getfield(observer, :nx)
+        return @view getfield(observer, :estimator).x[1:nx]
+    elseif name === :d
+        nx = getfield(observer, :nx)
+        ndo = getfield(observer, :nd_offsetfree)
+        return @view getfield(observer, :estimator).x[nx+1:nx+ndo]
+    elseif name === :K
+        return getfield(observer, :estimator).K
+    elseif name in fieldnames(OffsetFreeObserver)
+        return getfield(observer, name)
+    elseif name in fieldnames(KalmanFilter)
+        return getproperty(getfield(observer, :estimator), name)
+    else
+        return getfield(observer, name)
+    end
+end
+
+get_estimated_disturbance(::KalmanFilter) = zeros(0)
+get_estimated_disturbance(observer::OffsetFreeObserver) = collect(observer.d)
+
 function KalmanFilter(F,G,C;Gd=nothing,Dd=nothing,f_offset=nothing, h_offset=nothing, x0=nothing, Q=nothing,R=nothing)
     # Solve equation 
     ny,nx = size(C) 
@@ -21,7 +60,7 @@ function KalmanFilter(F,G,C;Gd=nothing,Dd=nothing,f_offset=nothing, h_offset=not
     h_offset = isnothing(h_offset) ? zeros(ny) : h_offset;
     x0 = isnothing(x0) ? zeros(nx) : x0;
     Q = isnothing(Q) ? Matrix{Float64}(I,nx,nx) : matrixify(Q,nx);
-    R = isnothing(R) ? Matrix{Float64}(I,ny,ny) : matrixify(R,nu);
+    R = isnothing(R) ? Matrix{Float64}(I,ny,ny) : matrixify(R,ny);
 
     P,_ = ared(F',C',R,Q);
     K = P*C'/(C*P*C'+R) 
@@ -32,16 +71,54 @@ end
 function set_state!(kf::KalmanFilter,x)
     kf.x .= x
 end
+function set_state!(observer::OffsetFreeObserver, x, d0=nothing)
+    xaug = observer.estimator.x
+    if length(x) == length(xaug)
+        xaug .= x
+    elseif length(x) == observer.nx
+        xaug[1:observer.nx] .= x
+        if isnothing(d0)
+            xaug[observer.nx+1:end] .= 0
+        else
+            length(d0) == observer.nd_offsetfree || throw(ArgumentError("Offset-free disturbance estimate must have length $(observer.nd_offsetfree)"))
+            xaug[observer.nx+1:end] .= d0
+        end
+    else
+        throw(ArgumentError("Observer state must have length $(observer.nx) or $(length(xaug))"))
+    end
+    return observer.x
+end
+
+function get_measured_disturbance(observer::OffsetFreeObserver, d)
+    ndm = observer.nd_measured
[…]
