+++ src/setup.jl
@@ -328,6 +328,153 @@ function set_state_observer!(mpc::Union{MPC,ExplicitMPC};
     mpc.state_observer = KalmanFilter(F,G,C;Gd,Dd,f_offset,h_offset,Q,R,x0)
 end
 
+function normalize_offset_free_method(method::Symbol)
+    aliases = Dict(
+        :state => :state_disturbance,
+        :state_disturbance => :state_disturbance,
+        :velocity => :velocity,
+        :output => :output_disturbance,
+        :output_disturbance => :output_disturbance,
+        :general => :general,
+    )
+    haskey(aliases, method) || throw(ArgumentError("Unknown offset-free method $method"))
+    return aliases[method]
+end
+
+function rebuild_model(model::Model, Gd, Dd, disturbance_labels)
+    labels = Labels(model.labels.x, model.labels.u, model.labels.y, Symbol.(disturbance_labels))
+    nd = size(Gd, 2)
+    return Model(model.F, model.G, float(Gd), model.f_offset, model.xo, model.uo,
+                 model.wmin, model.wmax, model.C, float(Dd), model.h_offset,
+                 model.true_dynamics, model.true_h,
+                 model.nx, model.nu, model.ny, nd, model.Ts, labels)
+end
+
+function strip_offset_free_model(model::Model, nd_measured::Int)
+    return rebuild_model(model, model.Gd[:,1:nd_measured], model.Dd[:,1:nd_measured], model.labels.d[1:nd_measured])
+end
+
+function append_offset_free_model(model::Model, Bd, Cd, disturbance_labels)
+    return rebuild_model(model, [model.Gd Bd], [model.Dd Cd], [model.labels.d; disturbance_labels])
+end
+
+function default_offset_free_labels(method::Symbol, nd::Int)
+    prefix = method == :output_disturbance ? "yoff" : "dof"
+    return Symbol.(prefix .* string.(1:nd))
+end
+
+function nominal_observer_gain(F, C; Q=nothing, R=nothing)
+    nx, ny = size(F, 1), size(C, 1)
+    return KalmanFilter(F, zeros(nx, ny), C; Q, R).K
+end
+
+function validate_offset_free_model(F, C, Bd, Cd)
+    nx = size(F, 1)
+    nd = size(Bd, 2)
+    ny = size(C, 1)
+    size(Bd, 1) == nx || throw(ArgumentError("Bd must have $nx rows"))
+    size(Cd) == (ny, nd) || throw(ArgumentError("Cd must have size ($ny, $nd)"))
+    rank([F - Matrix{Float64}(I, nx, nx) Bd; C Cd]) == nx + nd ||
+        throw(ArgumentError("Offset-free disturbance model violates rank([F-I Bd; C Cd]) = nx + nd"))
+end
+
+function build_offset_free_observer(model::Model, nd_measured::Int, method::Symbol;
+        Q=nothing, R=nothing, K=nothing, Bd=nothing, Cd=nothing,
+        Kx=nothing, Kd=nothing, x0=nothing, d0=nothing)
+
+    F, G, C = model.F, model.G, model.C
+    method = normalize_offset_free_method(method)
+    nx, ny = model.nx, model.ny
+
+    if method == :state_disturbance || method == :velocity
+        K = isnothing(K) ? nominal_observer_gain(F, C; Q, R) : float(K)
+        size(K) == (nx, ny) || throw(ArgumentError("K must have size ($nx, $ny)"))
+        Bd = K
+        Cd = Matrix{Float64}(I, ny, ny) - C*K
+        Kx = K
+        Kd = Matrix{Float64}(I, ny, ny)
+    elseif method == :output_disturbance
[…]
