### 2026-04-15 9530474
Add code generation for offset-free tracking
