### 2026-04-15 d21328d
Offset-free tracking  (#99)
* Initial commit for supporting offset-free observer

* Add code generation for offset-free tracking

* Improve docs
