+++ test/runtests.jl
@@ -643,6 +643,84 @@ Random.seed!(1234)
             @test norm(x-xref2) < 1e-9
         end
     end
+    @testset "Offset-free observer" begin
+        F,G = [1 0.1; 0 1], [0.005;0.1;;]
+        C = [1.0 0]
+        dynamics = (x,u,d) -> F*x + G*u + [0.01;0.0]
+
+        nominal = LinearMPC.MPC(F,G;Ts=0.1,C=C,Np=20)
+        set_objective!(nominal;Q=[1.0],R=[0.0],Rr=[0.1])
+        set_bounds!(nominal;umin=[-1.0],umax=[1.0])
+        sim_nominal = Simulation(dynamics, nominal; x0=zeros(2), N=100, r=[0.5])
+
+        tracked = LinearMPC.MPC(F,G;Ts=0.1,C=C,Np=20)
+        set_objective!(tracked;Q=[1.0],R=[0.0],Rr=[0.1])
+        set_bounds!(tracked;umin=[-1.0],umax=[1.0])
+        obs = LinearMPC.set_offset_free_observer!(tracked; method=:velocity,
+                                                  Q=[1e-3,1e-3], R=[1e-4])
+        sim_tracked = Simulation(dynamics, tracked; x0=zeros(2), N=100, r=[0.5])
+
+        @test obs.formulation == :velocity
+        @test size(tracked.model.Gd) == (2,1)
+        @test size(tracked.model.Dd) == (1,1)
+        @test length(LinearMPC.get_estimated_disturbance(tracked)) == 1
+        @test abs(sim_nominal.xs[1,end] - 0.5) > 5e-2
+        @test abs(sim_tracked.xs[1,end] - 0.5) < 1e-3
+
+        srcdir = tempname()
+        LinearMPC.codegen(tracked, dir=srcdir)
+        src = [f for f in readdir(srcdir) if last(f,1) == "c"]
+
+        if !isnothing(Sys.which("gcc"))
+            testlib = "mpctest."* Base.Libc.Libdl.dlext
+            run(Cmd(`gcc -lm -fPIC -O3 -msse3 -xc -shared -o $testlib $src`; dir=srcdir))
+
+            xaug = randn(3)
+            uin = randn(1)
+            y = randn(1)
+            ref = [0.5]
+
+            LinearMPC.set_state!(tracked.state_observer, xaug)
+            xref_pred = copy(LinearMPC.predict!(tracked.state_observer.estimator, uin))
+            xpred = copy(xaug)
+
+            global templib = joinpath(srcdir, testlib)
+            ccall(("mpc_predict_state", templib), Cvoid,
+                  (Ptr{Cdouble}, Ptr{Cdouble}, Ptr{Cdouble}), xpred, uin, C_NULL)
+            @test norm(xpred-xref_pred) < 1e-9
+
+            LinearMPC.set_state!(tracked.state_observer, xaug)
+            xref_corr = copy(LinearMPC.correct!(tracked.state_observer.estimator, y))
+            xcorr = copy(xaug)
+            ccall(("mpc_correct_state", templib), Cvoid,
+                  (Ptr{Cdouble}, Ptr{Cdouble}, Ptr{Cdouble}), xcorr, y, C_NULL)
+            @test norm(xcorr-xref_corr) < 1e-9
+
+            LinearMPC.set_state!(tracked.state_observer, xaug)
+            tracked.uprev .= 0
+            uref = compute_control(tracked, tracked.state_observer.x; r=ref)
+            uc = zeros(1)
+            ccall(("mpc_compute_control_observer", templib), Cint,
[…]
