### 2026-04-20 10b7df4
First attempt to include disturbance preview
