+++ test/runtests.jl
@@ -332,6 +332,81 @@ Random.seed!(1234)
         @test_nowarn compute_control(mpc, [1.0, 0.0]; r=[0.0 0.0; 0.0 0.0])  # Correct matrix
     end
 
+    @testset "Disturbance Preview" begin
+        A = [1.0 1.0; 0.0 1.0]
+        B = [0.0; 1.0]
+        Gd = [0.0; 1.0]
+        C = [1.0 0.0]
+        mpc = LinearMPC.MPC(A, B; Gd, C, Np=5, Nc=5)
+        set_bounds!(mpc; umin=[-0.5], umax=[0.5])
+        set_objective!(mpc; Q=[10.0], R=[0.1])
+
+        @test mpc.settings.disturbance_preview == false
+        control_standard = compute_control(mpc, [0.0, 0.0]; d=[0.0])
+        @test length(control_standard) == 1
+
+        mpc.settings.disturbance_preview = true
+        setup!(mpc)
+
+        control_single = compute_control(mpc, [0.0, 0.0]; d=[0.0])
+        d_traj = [0.0 1.0 1.0 1.0 1.0]
+        control_preview = compute_control(mpc, [0.0, 0.0]; d=d_traj)
+        @test length(control_preview) == 1
+
+        nx, nr, nd, nuprev = LinearMPC.get_parameter_dims(mpc)
+        @test nx == 2
+        @test nr == 1
+        @test nd == 5
+        @test nuprev == 0
+
+        @test control_preview[1] < control_single[1] - 1e-2
+        @test norm(control_preview - control_single) > 1e-2
+
+        empc = LinearMPC.ExplicitMPC(mpc; range=LinearMPC.ParameterRange(mpc), build_tree=true)
+        control_explicit = compute_control(empc, [0.0, 0.0]; d=d_traj)
+        @test control_explicit ≈ control_preview atol=1e-10
+    end
+
+    @testset "Disturbance Preview Simulation" begin
+        A = [1.0 1.0; 0.0 1.0]
+        B = [0.0; 1.0]
+        Gd = [0.0; 1.0]
+        C = [1.0 0.0]
+        mpc = LinearMPC.MPC(A, B; Gd, C, Np=5, Nc=5)
+        set_bounds!(mpc; umin=[-0.5], umax=[0.5])
+        set_objective!(mpc; Q=[10.0], R=[0.1])
+
+        N_sim = 20
+        d_traj = [zeros(1, 8) ones(1, 12)]
+
+        mpc.settings.disturbance_preview = true
+        setup!(mpc)
+        sim_preview = LinearMPC.Simulation(mpc; x0=[0.0, 0.0], N=N_sim, d=d_traj)
+        empc = LinearMPC.ExplicitMPC(mpc; range=LinearMPC.ParameterRange(mpc), build_tree=true)
+        sim_explicit_preview = LinearMPC.Simulation(empc; x0=[0.0, 0.0], N=N_sim, d=d_traj)
+
+        mpc.settings.disturbance_preview = false
+        setup!(mpc)
+        sim_no_preview = LinearMPC.Simulation(mpc; x0=[0.0, 0.0], N=N_sim, d=d_traj)
+
+        @test norm(sim_preview.us - sim_no_preview.us) > 1e-2
+        @test norm(sim_preview.ys - sim_explicit_preview.ys) < 1e-6
+        @test norm(sim_preview.ys) / norm(sim_no_preview.ys) < 0.9
+    end
+
+    @testset "Disturbance Preview Error Handling" begin
[…]
