### 2026-04-20 60e0e3f
Add test for disturbance preview codegen
