### 2026-04-20 b1d5384
Add disturbance preview coverage tests and fix ref_preview_cost dimension bug
Agent-Logs-Url: https://github.com/darnstrom/LinearMPC.jl/sessions/017a5318-047f-4791-86d2-78a04ed6bc6b

Co-authored-by: darnstrom <55484604+darnstrom@users.noreply.github.com>
