+++ test/runtests.jl
@@ -407,6 +407,186 @@ Random.seed!(1234)
         @test_nowarn compute_control(mpc, [0.0]; d=[0.0 1.0])
     end
 
+    @testset "Disturbance Preview Multiple Disturbances" begin
+        # Test with nd > 1: two independent disturbances
+        A = [0.9 0.0; 0.0 0.8]
+        B = [1.0; 0.0]
+        Gd = [1.0 0.0; 0.0 1.0]
+        C = [1.0 0.0; 0.0 1.0]
+        mpc = LinearMPC.MPC(A, B; Gd, C, Np=4, Nc=4)
+        set_bounds!(mpc; umin=[-2.0], umax=[2.0])
+        set_objective!(mpc; Q=[1.0, 1.0], R=[0.1])
+        mpc.settings.disturbance_preview = true
+        setup!(mpc)
+
+        # nd should be nd_base * Np = 2 * 4 = 8
+        nx, nr, nd, nuprev, nl = LinearMPC.get_parameter_dims(mpc)
+        @test nd == 2 * 4
+
+        # Constant disturbance vector should be tiled into preview
+        d_const = [0.5, -0.3]
+        u_const = compute_control(mpc, [0.0, 0.0]; d=d_const)
+        @test length(u_const) == 1
+
+        # Disturbance trajectory matrix (nd_base × Np)
+        d_traj = [0.0 0.5 1.0 1.0; 0.0 0.0 -0.5 -0.5]
+        u_traj = compute_control(mpc, [0.0, 0.0]; d=d_traj)
+        @test length(u_traj) == 1
+
+        # Different disturbance trajectories should yield different controls
+        @test norm(u_traj - u_const) > 1e-5
+    end
+
+    @testset "Disturbance Preview + Reference Preview Combined" begin
+        A = [1.0 1.0; 0.0 1.0]
+        B = [0.0; 1.0]
+        Gd = [0.5; 0.0]
+        C = [1.0 0.0]
+        mpc = LinearMPC.MPC(A, B; Gd, C, Np=5, Nc=5)
+        set_bounds!(mpc; umin=[-2.0], umax=[2.0])
+        set_objective!(mpc; Q=[5.0], R=[0.1])
+        mpc.settings.reference_preview = true
+        mpc.settings.disturbance_preview = true
+        setup!(mpc)
+
+        nx, nr, nd, nuprev, nl = LinearMPC.get_parameter_dims(mpc)
+        @test nx == 2
+        @test nr == 1 * 5   # ny × Np
+        @test nd == 1 * 5   # nd_base × Np
+
+        x = [0.0, 0.0]
+        r_traj = [0.0 0.5 1.0 1.0 1.0]
+        d_traj = [0.0 0.2 0.4 0.4 0.4]
+        u_both = compute_control(mpc, x; r=r_traj, d=d_traj)
+        @test length(u_both) == 1
+
+        # Without any preview for comparison
+        mpc.settings.reference_preview = false
+        mpc.settings.disturbance_preview = false
+        setup!(mpc)
+        u_neither = compute_control(mpc, x; r=[0.0], d=[0.0])
+        @test norm(u_both - u_neither) > 1e-4
+
+        # Restore both previews and verify explicit MPC matches
+        mpc.settings.reference_preview = true
+        mpc.settings.disturbance_preview = true
+        setup!(mpc)
+        empc = LinearMPC.ExplicitMPC(mpc; range=LinearMPC.ParameterRange(mpc), build_tree=true)
+        u_explicit = compute_control(empc, x; r=r_traj, d=d_traj)
+        @test u_explicit ≈ u_both atol=1e-10
+    end
+
[…]
