### 2026-04-22 eea3417
Add support for disturbance preview (#100)
* First attempt to include disturbance preview

* Add test for disturbance preview codegen

* Add disturbance preview coverage tests and fix ref_preview_cost dimension bug

Agent-Logs-Url: https://github.com/darnstrom/LinearMPC.jl/sessions/017a5318-047f-4791-86d2-78a04ed6bc6b

Co-authored-by: darnstrom <55484604+darnstrom@users.noreply.github.com>

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
