+++ src/mpc2mpqp.jl
@@ -45,22 +45,87 @@ function state_predictor(F,G,Np,Nc)
     return Φ::Matrix{Float64},Γ::Matrix{Float64}
 end
 
+function disturbance_predictor(F,E,Np)
+    nx,nd = size(E)
+    Ψ = zeros((Np+1)*nx, Np*nd)
+    for k in 1:Np
+        rows = k*nx+1:(k+1)*nx
+        prev = (k-1)*nx+1:k*nx
+        Ψ[rows, :] .= F*Ψ[prev, :]
+        Ψ[rows, (k-1)*nd+1:k*nd] .+= E
+    end
+    return Ψ::Matrix{Float64}
+end
+
+function disturbance_preview_predictor(mpc::MPC, F)
+    nd = mpc.model.nd
+    nxe = size(F,1)
+    nd == 0 && return zeros((mpc.Np+1)*nxe, 0)
+    E = [mpc.model.Gd; zeros(nxe-mpc.model.nx, nd)]
+    return disturbance_predictor(F, E, mpc.Np)
+end
+
+take_cols(W, start, len) = len == 0 ? zeros(size(W,1), 0) : W[:, start:start+len-1]
+
+function insert_preview_parameter_blocks(mpc::MPC, W, Wr, Wd)
+    nx = mpc.model.nx
+    nr_base = mpc.settings.reference_preview ? 0 : mpc.nr
+    nd_base = mpc.settings.disturbance_preview ? 0 : mpc.nd
+
+    idx = nx + nr_base + nd_base
+    parts = Matrix{Float64}[take_cols(W, 1, nx)]
+
+    if mpc.settings.reference_preview && !isempty(Wr)
+        push!(parts, Wr)
+    else
+        push!(parts, take_cols(W, nx+1, nr_base))
+    end
+
+    if mpc.settings.disturbance_preview && !isempty(Wd)
+        push!(parts, Wd)
+    else
+        push!(parts, take_cols(W, nx+nr_base+1, nd_base))
+    end
+
+    push!(parts, take_cols(W, idx+1, size(W,2)-idx))
+    return hcat(parts...)
+end
+
+function disturbance_preview_direct(A, ks, Np, nd)
+    m = size(A,1)
+    W = zeros(m*length(ks), nd*Np)
+    (isempty(A) || nd == 0) && return W
+
+    for (i,k) in enumerate(ks)
+        if 2 <= k <= Np+1
+            rows = (i-1)*m+1:i*m
+            cols = (k-2)*nd+1:(k-1)*nd
+            W[rows, cols] .= -A
+        end
+    end
+    return W
+end
+
 function get_parameter_dims(mpc::MPC)
     # Use stored values if QP is set up, otherwise compute from settings
     # This ensures consistency between the QP and parameter vector at runtime
     if mpc.mpqp_issetup
-        return mpc.model.nx, mpc.nr, mpc.model.nd, mpc.nuprev, mpc.nl
+        return mpc.model.nx, mpc.nr, mpc.nd, mpc.nuprev, mpc.nl
     end
     nr = mpc.settings.reference_tracking ?  mpc.model.ny : 0
     if mpc.settings.reference_preview && !mpc.settings.reference_condensation && nr > 0
         nr = nr * mpc.Np  # Reference preview uses Np time steps
     end
+    nd = mpc.model.nd
+    if mpc.settings.disturbance_preview && nd > 0
+        nd = nd * mpc.Np
+    end
     nuprev = !iszero(mpc.weights.Rr) || any(!iszero(c.Aup) for c in mpc.constraints) ? mpc.model.nu : 0
[…]
