### 2026-04-23 565ce9e
Minor editing of paper (#101)
