### 2026-04-23 3c86034
Trying to add support for generalized parameters
