+++ src/observer.jl
@@ -154,8 +154,12 @@ function codegen(observer::OffsetFreeObserver, mpc::Union{MPC,ExplicitMPC}, fh, 
     @printf(fh, "#define N_OFFSET_FREE_DISTURBANCE %d\n", observer.nd_offsetfree)
     @printf(fh, "void mpc_get_estimated_state(c_float* state, c_float* observer_state);\n")
     @printf(fh, "void mpc_get_estimated_disturbance(c_float* disturbance, c_float* observer_state, c_float* measured_disturbance);\n")
-    if mpc.nl > 0
+    if mpc.nl > 0 && mpc.np > 0
+        @printf(fh, "int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_float* reference, c_float* measured_disturbance, c_float* linear_cost, c_float* affine_parameter);\n")
+    elseif mpc.nl > 0
         @printf(fh, "int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_float* reference, c_float* measured_disturbance, c_float* linear_cost);\n")
+    elseif mpc.np > 0
+        @printf(fh, "int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_float* reference, c_float* measured_disturbance, c_float* affine_parameter);\n")
     else
         @printf(fh, "int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_float* reference, c_float* measured_disturbance);\n")
     end
@@ -173,7 +177,17 @@ void mpc_get_estimated_disturbance(c_float* disturbance, c_float* observer_state
 }
 """)
 
-    if mpc.nl > 0
+    if mpc.nl > 0 && mpc.np > 0
+        write(fsrc, """
+int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_float* reference, c_float* measured_disturbance, c_float* linear_cost, c_float* affine_parameter){
+    c_float state[N_STATE];
+    c_float disturbance[N_DISTURBANCE];
+    mpc_get_estimated_state(state, observer_state);
+    mpc_get_estimated_disturbance(disturbance, observer_state, measured_disturbance);
+    return mpc_compute_control(control, state, reference, disturbance, linear_cost, affine_parameter);
+}
+""")
+    elseif mpc.nl > 0
         write(fsrc, """
 int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_float* reference, c_float* measured_disturbance, c_float* linear_cost){
     c_float state[N_STATE];
@@ -181,6 +195,16 @@ int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_fl
     mpc_get_estimated_state(state, observer_state);
     mpc_get_estimated_disturbance(disturbance, observer_state, measured_disturbance);
     return mpc_compute_control(control, state, reference, disturbance, linear_cost);
+}
+""")
+    elseif mpc.np > 0
+        write(fsrc, """
+int mpc_compute_control_observer(c_float* control, c_float* observer_state, c_float* reference, c_float* measured_disturbance, c_float* affine_parameter){
+    c_float state[N_STATE];
+    c_float disturbance[N_DISTURBANCE];
+    mpc_get_estimated_state(state, observer_state);
+    mpc_get_estimated_disturbance(disturbance, observer_state, measured_disturbance);
+    return mpc_compute_control(control, state, reference, disturbance, affine_parameter);
 }
 """)
     else
