+++ codegen/mpc_update_qp.c
@@ -26,9 +26,15 @@ void mpc_get_solution(c_float* th, c_float* control, c_float* xstar){
 #include "bnb.h"
 #endif
 
-#if N_LINEAR_COST > 0
+#if N_LINEAR_COST > 0 && N_AFFINE_PARAMETER > 0
+int mpc_compute_control(c_float* control, c_float* state, c_float* reference, c_float* disturbance, c_float* linear_cost, c_float* affine_parameter){
+    mpc_update_parameter(mpc_parameter, control, state, reference, disturbance, linear_cost, affine_parameter);
+#elif N_LINEAR_COST > 0
 int mpc_compute_control(c_float* control, c_float* state, c_float* reference, c_float* disturbance, c_float* linear_cost){
     mpc_update_parameter(mpc_parameter, control, state, reference, disturbance, linear_cost);
+#elif N_AFFINE_PARAMETER > 0
+int mpc_compute_control(c_float* control, c_float* state, c_float* reference, c_float* disturbance, c_float* affine_parameter){
+    mpc_update_parameter(mpc_parameter, control, state, reference, disturbance, affine_parameter);
 #else
 int mpc_compute_control(c_float* control, c_float* state, c_float* reference, c_float* disturbance){
     mpc_update_parameter(mpc_parameter, control, state, reference, disturbance);

[…]
