### 2026-04-24 52c7ccc
Remove linear cost, since it is covered by generalized parameters
