+++ codegen/mpc_update_parameter.c
@@ -1,8 +1,4 @@
-#if N_LINEAR_COST > 0 && N_AFFINE_PARAMETER > 0
-void mpc_update_parameter(c_float* parameter, c_float* control, c_float* state, c_float* reference, c_float* disturbance, c_float* linear_cost, c_float* affine_parameter){
-#elif N_LINEAR_COST > 0
-void mpc_update_parameter(c_float* parameter, c_float* control, c_float* state, c_float* reference, c_float* disturbance, c_float* linear_cost){
-#elif N_AFFINE_PARAMETER > 0
+#if N_AFFINE_PARAMETER > 0
 void mpc_update_parameter(c_float* parameter, c_float* control, c_float* state, c_float* reference, c_float* disturbance, c_float* affine_parameter){
 #else
 void mpc_update_parameter(c_float* parameter, c_float* control, c_float* state, c_float* reference, c_float* disturbance){
@@ -27,26 +23,6 @@ void mpc_update_parameter(c_float* parameter, c_float* control, c_float* state, 
     for(j=0;j<N_DISTURBANCE_BASE;i++, j++) parameter[i] = disturbance[j];
 #endif
     for(j=0;j<N_CONTROL_PREV;i++, j++) parameter[i] = control[j];
-#if N_LINEAR_COST > 0
-#ifdef N_MOVE_BLOCKS
-    // Average linear cost over move blocks
-    // linear_cost is (N_CONTROL x N_PREDICTION_HORIZON), column-major
-    int block_offset = 0;
-    for(int b = 0; b < N_MOVE_BLOCKS; b++) {
-        int block_size = move_blocks[b];
-        for(int u = 0; u < N_CONTROL; u++) {
-            c_float sum = 0.0;
-            for(int k = 0; k < block_size; k++) {
-                sum += linear_cost[u + (block_offset + k) * N_CONTROL];
-            }
-            parameter[i++] = sum / block_size;
-        }
-        block_offset += block_size;
-    }
-#else
-    for(j=0;j<N_LINEAR_COST;i++, j++) parameter[i] = linear_cost[j];
-#endif
-#endif
 #if N_AFFINE_PARAMETER > 0
     for(j=0;j<N_AFFINE_PARAMETER;i++, j++) parameter[i] = affine_parameter[j];
 #endif
