+++ src/mpc2mpqp.jl
@@ -126,7 +126,7 @@ function get_parameter_dims(mpc::MPC)
     # Use stored values if QP is set up, otherwise compute from settings
     # This ensures consistency between the QP and parameter vector at runtime
     if mpc.mpqp_issetup
-        return mpc.model.nx, mpc.nr, mpc.nd, mpc.nuprev, mpc.nl, mpc.np
+        return mpc.model.nx, mpc.nr, mpc.nd, mpc.nuprev, mpc.np
     end
     nr = mpc.settings.reference_tracking ?  mpc.model.ny : 0
     if mpc.settings.reference_preview && !mpc.settings.reference_condensation && nr > 0
@@ -137,13 +137,12 @@ function get_parameter_dims(mpc::MPC)
         nd = nd * mpc.Np
     end
     nuprev = !iszero(mpc.weights.Rr) || any(!iszero(c.Aup) for c in mpc.constraints) ? mpc.model.nu : 0
-    nl = mpc.settings.linear_cost ? mpc.model.nu * mpc.Nc : 0
     np = get_affine_parameter_base_dim(mpc) * mpc.Np
-    return mpc.model.nx,nr,nd,nuprev,nl,np
+    return mpc.model.nx,nr,nd,nuprev,np
 end
 
 function get_parameter_names(mpc::Union{MPC,ExplicitMPC})
-    nx,nr,nd,nuprev,nl,np = get_parameter_dims(mpc)
+    nx,nr,nd,nuprev,np = get_parameter_dims(mpc)
     names = copy(mpc.model.labels.x)
     if nr > 0
         if mpc.settings.reference_preview && !mpc.settings.reference_condensation
@@ -165,11 +164,6 @@ function get_parameter_names(mpc::Union{MPC,ExplicitMPC})
         end
     end
     nuprev>0 && push!(names,Symbol.(string.(mpc.model.labels.u).*"p")...)
-    if nl > 0
-        for k in 0:mpc.Nc-1
-            push!(names, Symbol.(string.(mpc.model.labels.u).*"l_$k")...)
-        end
-    end
     if np > 0
         np_base = get_affine_parameter_base_dim(mpc)
         for k in 0:mpc.Np-1, i in 1:np_base
@@ -183,8 +177,8 @@ end
 # correspoding to  lb<=u_i<=ub for i ∈ {1,2,...,Nc}
 function create_controlbounds(mpc::MPC, F, Γ, Φ)
     nu,nx,Nb = mpc.model.nu, mpc.model.nx, mpc.Nc
-    _,_,_,_,nl,np = get_parameter_dims(mpc)
-    nxe = sum(get_parameter_dims(mpc))-nl-np
+    _,_,_,_,np = get_parameter_dims(mpc)
+    nxe = sum(get_parameter_dims(mpc))-np
     mpc.settings.reference_preview && (nxe-=mpc.nr) # reference not part of extended state
     mpc.settings.disturbance_preview && (nxe-=mpc.nd) # disturbance not part of extended state
     !iszero(mpc.model.f_offset) && (nxe+=1); # constant offset in dynamics
@@ -218,7 +212,6 @@ function create_controlbounds(mpc::MPC, F, Γ, Φ)
         zeros(size(W,1), 0)
     end
     W = insert_preview_parameter_blocks(mpc, W, Wr, Wd)
-    nl > 0 && (W = [W zeros(size(W,1), nl)])
     np > 0 && (W = [W zeros(size(W,1), np)])
     return A,ub,lb, W
 end
[…]
