### 2026-04-24 05caff0
Also add affine cost term for states
