+++ src/types.jl
@@ -25,19 +25,24 @@ struct MPCWeights
     S::Matrix{Float64}
     Qf::Matrix{Float64}
     Qfx::Matrix{Float64}
+    Ex::Matrix{Float64}
+    ex::Vector{Float64}
     E::Matrix{Float64}
     e::Vector{Float64}
 end
 
 function MPCWeights(nu,nx,nr)
     return MPCWeights(Matrix{Float64}(I,nr,nr),Matrix{Float64}(I,nu,nu),zeros(nu,nu),
-                      zeros(nx,nu),zeros(nr,nr),zeros(nx,nx),zeros(nu,0),zeros(nu))
+                      zeros(nx,nu),zeros(nr,nr),zeros(nx,nx),zeros(nx,0),zeros(nx),zeros(nu,0),zeros(nu))
 end
 
 function MPCWeights(Q::AbstractArray,R::AbstractArray,Rr::AbstractArray=zeros(size(R));
-        S = zeros(0,0), Qf = zeros(0,0), Qfx = zeros(0,0), E = zeros(size(R, 1), 0), e = zeros(size(R, 1)))
+        S = zeros(0,0), Qf = zeros(0,0), Qfx = zeros(0,0),
+        Ex = zeros(size(Q, 1), 0), ex = zeros(size(Q, 1)),
+        E = zeros(size(R, 1), 0), e = zeros(size(R, 1)))
     Qf = isempty(Qf) ? copy(Q) : Qf 
-    return MPCWeights(matrixify(Q),matrixify(R),matrixify(Rr),float(S),matrixify(Qf),matrixify(Qfx),float(E),float(e))
+    return MPCWeights(matrixify(Q),matrixify(R),matrixify(Rr),float(S),matrixify(Qf),matrixify(Qfx),
+                      float(Ex),float(ex),float(E),float(e))
 end
 
 """
