### 2026-04-24 1212cb4
E -> Eu, e -> eu
