### 2026-04-25 4ae1c57
Make parameter preview optional
