### 2026-04-26 1291d81
Generalized parameters (#102)
* Trying to add support for generalized parameters

* Remove linear cost, since it is covered by generalized parameters

* Also add affine cost term for states

* E -> Eu, e -> eu

* Make parameter preview optional
