+++ src/utils.jl
@@ -1,5 +1,5 @@
 """
-    compute_control(mpc,x;r,d,uprev,l)
+    compute_control(mpc,x;r,d,uprev,p)
 
 For a given MPC `mpc` and state `x`, compute the optimal control action.
 
@@ -11,11 +11,10 @@ Optional arguments:
   - Vector of length `nd` for constant disturbance
   - Matrix of size `(nd, Np)` for disturbance preview (when `mpc.settings.disturbance_preview = true`)
 * `uprev` - previous control action
-* `l` - linear cost on control (requires `mpc.settings.linear_cost = true`). Can be:
-  - Vector of length `nu` for constant linear cost across horizon
-  - Matrix of size `(nu, Np)` for time-varying linear cost (mean computed per move block)
-  - Matrix of size `(nu, Nc)` for pre-blocked linear cost
-  - Nothing (default) for no linear cost
+* `p` - generalized parameter trajectory. Can be:
+  - Vector of length `np` for a constant parameter across the horizon
+  - Matrix of size `(np, Np)` for time-varying generalized parameters when `mpc.settings.parameter_preview = true`
+  - Nothing (default) for no generalized-parameter contribution
 
 All arguments default to zero.
 
@@ -33,16 +32,16 @@ u = compute_control(mpc, x; r=r_trajectory)
 d_trajectory = [0.0 0.2 0.4 0.4 0.4]  # nd × Np matrix
 u = compute_control(mpc, x; d=d_trajectory)
 
-# With linear cost (requires mpc.settings.linear_cost = true)
-u = compute_control(mpc, x; l=[1.0])  # Constant cost
+# Generalized parameter preview
+u = compute_control(mpc, x; p=[0.5, -0.25])
 
-# Time-varying linear cost
-l_trajectory = [1.0 0.5 0.0 -0.5 -1.0]  # nu × Np matrix
-u = compute_control(mpc, x; l=l_trajectory)
+# Generalized parameter preview over the horizon
+mpc.settings.parameter_preview = true
+u = compute_control(mpc, x; p=[0.5 0.25 0.0 -0.25 -0.5])
 ```
 """
-function compute_control(mpc::MPC,x;r=nothing,d=nothing,uprev=nothing,l=nothing, check=true)
-    θ = form_parameter(mpc,x,r,d,uprev,l)
+function compute_control(mpc::MPC,x;r=nothing,d=nothing,uprev=nothing,p=nothing, check=true)
+    θ = form_parameter(mpc,x,r,d,uprev,p)
     udaqp,fval,exitflag,info = solve(mpc,θ)
     check && @assert(exitflag>=1)
     # mpc.uprev = udaqp[1:mpc.model.nu]-mpc.K*θ[1:mpc.model.nx]
@@ -51,17 +50,17 @@ function compute_control(mpc::MPC,x;r=nothing,d=nothing,uprev=nothing,l=nothing,
     return copy(mpc.uprev)
 end
 
-function compute_control(empc::ExplicitMPC,x;r=nothing,d=nothing,uprev=nothing,l=nothing, check=true)
+function compute_control(empc::ExplicitMPC,x;r=nothing,d=nothing,uprev=nothing,p=nothing, check=true)
     if isnothing(empc.bst)
         @error "Need to build a binary search tree to evaluate control law"
     else
-        empc.uprev .= ParametricDAQP.evaluate(empc.bst,form_parameter(empc,x,r,d,uprev,l))
+        empc.uprev .= ParametricDAQP.evaluate(empc.bst,form_parameter(empc,x,r,d,uprev,p))
         return copy(empc.uprev) 
     end
 end
 
-function compute_control_trajectory(mpc::MPC,x;r=nothing,d=nothing,uprev=nothing,l=nothing, check=true)
[…]
