### 2026-04-28 53402ea
Fix new example interface
