+++ src/mpc_examples.jl
@@ -5,301 +5,378 @@ struct MPCExample
     scenarios::Vector{Scenario}
 end
 
+MPCExample(name::AbstractString, mpc::MPC, range::ParameterRange; scenarios=Scenario[]) =
+    MPCExample(String(name), mpc, range, Scenario[scenarios...])
+
+_normalize_example_id(s::Union{AbstractString,Symbol}) = replace(lowercase(String(s)), r"[\s_-]" => "")
+
+const _MPC_EXAMPLE_SPECS = (
+    invpend = (
+        name = "Inverted Pendulum on a Cart",
+        aliases = ("inv_pend", "invpend", "invpendcart"),
+        defaults = (Np = 50, Nc = 5),
+    ),
+    dcmotor = (
+        name = "DC Motor Position Control",
+        aliases = ("dc_motor", "dcmotor"),
+        defaults = (Np = 10, Nc = 2),
+    ),
+    aircraft = (
+        name = "Aircraft",
+        aliases = ("aircraft",),
+        defaults = (Np = 10, Nc = 2),
+    ),
+    chained = (
+        name = "Chained",
+        aliases = ("chained", "chained-firstorder"),
+        defaults = (Np = 10, Nc = 10),
+    ),
+    mass_spring = (
+        name = "Mass-Spring System",
+        aliases = ("mass-spring", "mass", "spring"),
+        defaults = (Np = 10, Nc = 10),
+    ),
+    nonlinear_demo = (
+        name = "Linearized Nonlinear Demo",
+        aliases = ("nonlinear", "nonlin"),
+        defaults = (Np = 5, Nc = 2),
+    ),
+    invpend_contact = (
+        name = "Colliding Inverted Pendulum on a Cart",
+        aliases = ("invpend_contact",),
+        defaults = (Np = 10, Nc = 10),
+    ),
+    ballplate = (
+        name = "Ball and Plate",
+        aliases = ("ball", "ballplate"),
+        defaults = (Np = 10, Nc = 2),
+    ),
+    crazyflie = (
+        name = "Quadcopter",
+        aliases = ("quad", "quadcopter", "crazyflie"),
+        defaults = (Np = 10, Nc = 10),
+    ),
+    satellite = (
+        name = "Satellite",
+        aliases = ("satellite",),
+        defaults = (Np = 20, Nc = 20),
+    ),
+)
+
+const _MPC_EXAMPLE_ALIASES = let aliases = Dict{String,Symbol}()
+    for (id, spec) in pairs(_MPC_EXAMPLE_SPECS)
+        aliases[_normalize_example_id(String(id))] = id
+        for alias in spec.aliases
+            aliases[_normalize_example_id(alias)] = id
+        end
+    end
+    aliases
+end
+
+_mpc_example_ids() = collect(propertynames(_MPC_EXAMPLE_SPECS))
+mpc_example_names() = sort!(String.(collect(propertynames(_MPC_EXAMPLE_SPECS))))
+
+function _resolve_example_id(s::Union{AbstractString,Symbol})
[…]
