### 2026-04-28 376505a
Restructuring MPC examples (#103)
* Start structuring MPC examples

* Fix new example interface

* No need to talk about examples in README yet.
