+++ src/mpc_examples.jl
@@ -1,410 +1,544 @@
-function mpc_examples(s, Np, Nc=Np;params = Dict(),settings=nothing)
-    if(s=="inv_pend"||s=="invpend" || s=="invpendcart")
-        # Inverted pendulum on a cart
-        M,m,g,l,damp = 1.0, 1.0, 9.81, 0.5, 10.0;
-        scale, Mm = 100,M+m
-
-        f=(x,u,d)->[x[2];
-                    (scale*u[1]-damp*x[2]-m*l*x[4]^2*sin(x[3])+m*g*sin(x[3])*cos(x[3]))/(M+m*sin(x[3])^2);
-                    x[4];
-                    (g*sin(x[3])+(scale*u[1]-damp*x[2]-m*l*x[4]^2*sin(x[3]))*cos(x[3])/Mm)/(l-m*l*cos(x[3])^2/Mm)];
-        h = (x,u,d) -> [x[1];x[3]] 
-
-        Ts = 0.01;
-        xo,uo = zeros(4),zeros(1)
-        model = LinearMPC.Model(f,h,xo,uo,Ts)
-
-        mpc = MPC(model;Nc,Np);
-
-        set_objective!(mpc;Q=[1.2^2,1],R=[0.0],Rr=[1.0])
-        set_bounds!(mpc; umin=[-2.0],umax=[2.0])
-
-        if(!isnothing(settings))
-            mpc.settings=settings
-        end
-
-        range = ParameterRange(mpc);
-        range.xmax[:] .=  20*ones(4)
-        range.xmin[:] .= -20*ones(4)
-        range.rmax[:] .= 20*ones(2)
-        range.rmin[:] .= -20*ones(2)
-        range.dmax[:] .= 20*ones(1)
-        range.dmin[:] .= -20*ones(1)
-
-    elseif(s=="dc_motor"||s=="dcmotor")
-        A = [0 1.0 0 0; -51.21 -1 2.56 0; 0 0 0 1; 128 0 -6.401 -10.2];
-        B = 440*[0;0;0;1.0;;];
-        C = [1 0 0 0;1280 0 -64.01 0];
-        Ts = 0.1;
-        tau = 78.5398;
-        C = C./[2*pi;2*tau]; # Scaling 
-
-
-        mpc = MPC(A,B,Ts;C,Np,Nc);
-
-        Q = [0.1^2, 0]; 
-        R = [0.0]
-        Rr = [0.1^2]
-        set_objective!(mpc;Q,R,Rr)
-
-        umin,umax = [-0.5],[0.5]
-        set_bounds!(mpc;umin,umax)
-
-        add_constraint!(mpc,Ax=C[2:2,:],lb = [-0.5] ,ub = [0.5], ks = 2:min(mpc.Nc+2,mpc.Np), soft=true)
-
-        if(isnothing(settings))
-            mpc.settings.reference_tracking=true;
-        else
-            mpc.settings=settings
-        end
-
-        range = ParameterRange(mpc);
-        range.xmax[:] = [ 4*pi  4*pi/Ts  4*pi*20  4*pi*20/Ts]
-        range.xmin[:] = -[ 4*pi  4*pi/Ts  4*pi*20  4*pi*20/Ts]
-        range.rmax[:] .= [5;0.5]
-        range.rmin[:] .=-[5;0.5]
-        range.umax[:] .= [0.5023]
-        range.umin[:] .= -[0.5023] 
-
-    elseif(s=="aircraft")
-        A = [-0.0151 -60.5651  0       -32.174;
-             -0.0001   -1.3411  0.9929   0    ;
-             0.00018  43.2541 -0.86939  0     ;
[…]
