+++ example/certify.jl
@@ -1,6 +1,7 @@
 using LinearMPC
-# Load inverted pendulum example and certify the iteration complexity 
-mpc,range = LinearMPC.mpc_examples("invpend")
+# Load inverted pendulum example and certify the iteration complexity
+example = LinearMPC.mpc_example("invpend")
+mpc,range = example.mpc, example.range
 result = LinearMPC.certify(mpc;range)
 
 using Plots
