+++ src/mpc_examples.jl
@@ -449,39 +449,69 @@ function _build_mpc_example(::Val{:ballplate}, Np, Nc; settings=nothing, kwargs.
 end
 
 function _build_mpc_example(::Val{:crazyflie}, Np, Nc; settings=nothing, kwargs...)
-    F = [1.0 0.0 0.0 0.0000000 0.0009810 0.0000000 0.0100000 0.0000000 0.0000000 0.0000000 0.0000016 0.0000000;
-         0.0 1.0 0.0 -0.0009810 0.0000000 0.0000000 0.0000000 0.0100000 0.0000000 -0.0000016 0.0000000 0.0000000;
-         0.0 0.0 1.0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0100000 0.0000000 0.0000000 0.0000000;
-         0.0 0.0 0.0 1.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0050000 0.0000000 0.0000000;
-         0.0 0.0 0.0 0.0000000 1.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0050000 0.0000000;
-         0.0 0.0 0.0 0.0000000 0.0000000 1.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0050000;
-         0.0 0.0 0.0 0.0000000 0.1962000 0.0000000 1.0000000 0.0000000 0.0000000 0.0000000 0.0004905 0.0000000;
-         0.0 0.0 0.0 -0.1962000 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 -0.0004905 0.0000000 0.0000000;
-         0.0 0.0 0.0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 0.0000000 0.0000000;
-         0.0 0.0 0.0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 0.0000000;
-         0.0 0.0 0.0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000;
-         0.0 0.0 0.0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 1.0000000]
-    G = [-0.0000011 0.0000012 0.0000011 -0.0000012;
-         0.0000011 0.0000012 -0.0000011 -0.0000012;
-         0.0002102 0.0002102 0.0002102 0.0002102;
-         -0.0068839 -0.0075809 0.0068916 0.0075732;
-         -0.0069177 0.0076070 0.0069392 -0.0076285;
-         0.0004937 -0.0001806 -0.0006961 0.0003830;
-         -0.0004524 0.0004975 0.0004538 -0.0004989;
-         0.0004502 0.0004958 -0.0004507 -0.0004953;
-         0.0420429 0.0420429 0.0420429 0.0420429;
-         -2.7535461 -3.0323404 2.7566264 3.0292601;
-         -2.7670702 3.0427842 2.7756950 -3.0514090;
-         0.1974771 -0.0722364 -0.2784376 0.1531969]
+    mass = get(kwargs, :mass, 0.035)
+    arm_length= get(kwargs, :arm_length, 0.046/1.414213562)
+    J = get(kwargs, :J,[1.66e-5 0.83e-6 0.72e-6; 
+                        0.83e-6 1.66e-5 1.8e-6; 
[…]
