### 2026-05-23 24f869c
Add rocket example
