### 2026-05-23 0f88f7d
Add rocket example (#105)
* Add rocket example

* Improve test coverage
