+++ src/setup.jl
@@ -94,7 +94,7 @@ Adds the constraints lb ≤ C x  ≤ ub for the time steps k ∈ ks
 function set_output_bounds!(mpc::MPC; ymin=zeros(0), ymax=zeros(0), ks = 2:mpc.Np, soft = true, binary=false, prio = 0)
     lb = !isempty(ymin) ? ymin-mpc.model.h_offset : zeros(0)
     ub = !isempty(ymax) ? ymax-mpc.model.h_offset : zeros(0)
-    add_constraint!(mpc, Ax = mpc.model.C, Ad = mpc.model.Dd, lb = lb, ub = ub; ks,soft,binary,prio)
+    add_constraint!(mpc, Ax = mpc.model.C, Au = mpc.model.D, Ad = mpc.model.Dd, lb = lb, ub = ub; ks,soft,binary,prio)
 end
 
 """
@@ -279,6 +279,132 @@ function set_binary_controls!(mpc,bin_ids,Nc_binary=-1)
     mpc.Nc_binary = Nc_binary
     mpc.mpqp_issetup = false
 end
+
+function require_mld(mpc::MPC)
+    isnothing(mpc.mld) && throw(ArgumentError("This controller was not created from an MLDModel"))
+    return mpc.mld
+end
+
+function set_auxiliary_bounds!(mpc::MPC; zmin=zeros(0), zmax=zeros(0))
+    mld = require_mld(mpc)
+    zmin = isempty(zmin) ? mld.zmin : float(zmin)
+    zmax = isempty(zmax) ? mld.zmax : float(zmax)
+    length(zmin) == mld.nz || throw(ArgumentError("zmin must have length $(mld.nz)"))
+    length(zmax) == mld.nz || throw(ArgumentError("zmax must have length $(mld.nz)"))
+    mld.zmin[:] = zmin
+    mld.zmax[:] = zmax
+    isempty(mpc.umin) || (mpc.umin[mld.ncontrols+mld.ndelta+1:end] .= zmin)
+    isempty(mpc.umax) || (mpc.umax[mld.ncontrols+mld.ndelta+1:end] .= zmax)
+    mpc.mpqp_issetup = false
+end
+
+function add_mld_constraint!(mpc::MPC;
+        Ax = nothing, Au = zeros(0,0), Aδ = zeros(0,0), Az = zeros(0,0),
+        Ar = zeros(0,0), Aw = zeros(0,0), Ad = zeros(0,0), Aup = zeros(0,0), Ap = zeros(0,0),
+        E1 = zeros(0,0), E2 = zeros(0,0), E3 = zeros(0,0), E4 = zeros(0,0), E5 = zeros(0),
+        ub = zeros(0), lb = zeros(0), ks = 1:mpc.Np, soft = false, binary = false, prio = 0)
+
+    mld = require_mld(mpc)
+    if !isempty(E5) || !isempty(E1) || !isempty(E2) || !isempty(E3) || !isempty(E4)
+        Ax = isempty(E4) ? zeros(length(E5), mpc.model.nx) : -float(E4)
+        Au = isempty(E1) ? zeros(length(E5), mld.ncontrols) : -float(E1)
+        Aδ = isempty(E2) ? zeros(length(E5), mld.ndelta) : float(E2)
+        Az = isempty(E3) ? zeros(length(E5), mld.nz) : float(E3)
+        ub = isempty(ub) ? float(E5) : ub
+        lb = isempty(lb) ? fill(-1e30, length(E5)) : lb
+    end
+
+    m = max(length(lb), length(ub))
+    if !isnothing(Ax)
[…]
