### 2026-05-23 b6925ed
Refine notation
