+++ src/model.jl
@@ -139,63 +139,66 @@ struct MLDModel
     ncontrols::Int
     ndelta::Int
     nz::Int
-    E1::Matrix{Float64}
-    E2::Matrix{Float64}
-    E3::Matrix{Float64}
-    E4::Matrix{Float64}
-    E5::Vector{Float64}
+    Eu::Matrix{Float64}
+    Edelta::Matrix{Float64}
+    Ez::Matrix{Float64}
+    Ex::Matrix{Float64}
+    be::Vector{Float64}
     zmin::Vector{Float64}
     zmax::Vector{Float64}
 end
 
-function MLDModel(F,B1,B2,B3;
-        C=zeros(0,0), D1=zeros(0,0), D2=zeros(0,0), D3=zeros(0,0),
-        E1=zeros(0,0), E2=zeros(0,0), E3=zeros(0,0), E4=zeros(0,0), E5=zeros(0),
-        B5=zeros(0), D5=zeros(0), Ts=-1.0, xo=zeros(0), uo=zeros(0),
+function MLDModel(F,Bu,Bdelta,Bz;
+        C=zeros(0,0), Du=zeros(0,0), Ddelta=zeros(0,0), Dz=zeros(0,0),
+        Eu=zeros(0,0), Edelta=zeros(0,0), Ez=zeros(0,0), Ex=zeros(0,0), be=zeros(0),
+        bx=zeros(0), by=zeros(0), Ts=-1.0, xo=zeros(0), uo=zeros(0),
         zmin=zeros(0), zmax=zeros(0), delta_labels=nothing, z_labels=nothing)
 
-    B1 = reshape(B1, size(B1,1), :)
-    B2 = reshape(B2, size(B2,1), :)
-    B3 = reshape(B3, size(B3,1), :)
+    Bu = reshape(Bu, size(Bu,1), :)
+    Bdelta = reshape(Bdelta, size(Bdelta,1), :)
+    Bz = reshape(Bz, size(Bz,1), :)
     nx = size(F, 1)
-    size(B1, 1) == nx || throw(ArgumentError("B1 must have $nx rows"))
-    size(B2, 1) == nx || throw(ArgumentError("B2 must have $nx rows"))
-    size(B3, 1) == nx || throw(ArgumentError("B3 must have $nx rows"))
+    size(Bu, 1) == nx || throw(ArgumentError("Bu must have $nx rows"))
+    size(Bdelta, 1) == nx || throw(ArgumentError("Bdelta must have $nx rows"))
+    size(Bz, 1) == nx || throw(ArgumentError("Bz must have $nx rows"))
 
-    ncontrols = size(B1, 2)
-    ndelta = size(B2, 2)
-    nz = size(B3, 2)
+    ncontrols = size(Bu, 2)
+    ndelta = size(Bdelta, 2)
+    nz = size(Bz, 2)
     ny = isempty(C) ? nx : size(C, 1)
 
-    D1 = isempty(D1) ? zeros(ny, ncontrols) : float(D1)
-    D2 = isempty(D2) ? zeros(ny, ndelta) : float(D2)
-    D3 = isempty(D3) ? zeros(ny, nz) : float(D3)
-    size(D1) == (ny, ncontrols) || throw(ArgumentError("D1 must have size ($ny, $ncontrols)"))
-    size(D2) == (ny, ndelta) || throw(ArgumentError("D2 must have size ($ny, $ndelta)"))
-    size(D3) == (ny, nz) || throw(ArgumentError("D3 must have size ($ny, $nz)"))
+    Du = isempty(Du) ? zeros(ny, ncontrols) : float(Du)
+    Ddelta = isempty(Ddelta) ? zeros(ny, ndelta) : float(Ddelta)
+    Dz = isempty(Dz) ? zeros(ny, nz) : float(Dz)
[…]
