+++ src/mpc_examples.jl
@@ -342,31 +342,13 @@ function _build_mpc_example(::Val{:invpend_contact}, Np, Nc; settings=nothing, k
          0 0 0;
          1 / mc 0 0;
          1 / (mc * l) -1 / (mp * l) 1 / (mp * l)]
-    B = [B zeros(4, 4)]
 
-    C, D = Matrix{Float64}(I, 4, 4), zeros(4, 1)
+    C = Matrix{Float64}(I, 4, 4)
     Ts = 0.05
     F, G = zoh(A, B, Ts)
 
-    mpc = MPC(F, G; C, Np, Nc)
-    Q = [1.0, 1, 1, 1]
-    R = [1.0; 1e-4 * ones(6)]
-    Rr = zeros(7)
-    Qf, ~ = ared(mpc.model.F, mpc.model.G[:, 1], mpc.weights.R[1:1, 1:1], mpc.weights.Q)
-    set_objective!(mpc; Q, R, Rr, Qf)
-    set_bounds!(mpc, umin = [-1.0; 0; zeros(4)], umax = [1.0; 1e30; 1e30; ones(4)])
-    set_binary_controls!(mpc, collect(4:7))
-
-    if isnothing(settings)
-        mpc.settings.reference_tracking = false
-    else
-        mpc.settings = settings
-    end
-
     uby = [d; pi / 10; 1; 1]
     lby = -uby
-    set_output_bounds!(mpc, ymin = lby, ymax = uby, ks = 2:mpc.Nc)
-
     δ2l, δ2u = -uby[1] + l * lby[2] - d, -lby[1] + l * uby[2] - d
     dotδ2l, dotδ2u = -uby[3] + l * lby[4], -lby[3] + l * uby[4]
     δ3l, δ3u = lby[1] - l * uby[2] - d, uby[1] - l * lby[2] - d
@@ -375,49 +357,66 @@ function _build_mpc_example(::Val{:invpend_contact}, Np, Nc; settings=nothing, k
     u2l, u2u = κ * δ2l + ν * dotδ2l, κ * δ2u + ν * dotδ2u
     u3l, u3u = κ * δ3l + ν * dotδ3l, κ * δ3u + ν * dotδ3u
 
-    Ax = [-1 l 0 0;
-          1 -l 0 0;
-          -κ κ*l -ν ν*l;
-          κ -κ*l ν -ν*l;
-          zeros(2, 4);
-          κ -κ*l ν -ν*l;
-          -κ κ*l -ν ν*l]
-    Au2 = [0 0 0 -δ2u 0 0 0;
-           0 0 0 -δ2l 0 0 0;
-           0 0 0 0 0 -u2u 0;
-           0 0 0 0 0 -u2l 0;
-           0 1 0 -u2u 0 0 0;
-           0 1 0 0 0 -u2u 0;
-           0 1 0 0 0 -u2l 0;
-           0 -1 0 u2u 0 0 0]
-    Au3 = [0 0 0 0 -δ3u 0 0;
-           0 0 0 0 -δ3l 0 0;
-           0 0 0 0 0 0 -u3u;
-           0 0 0 0 0 0 -u3l;
-           0 0 1 0 -u3u 0 0;
-           0 0 1 0 0 0 -u3u;
-           0 0 1 0 0 0 -u3l;
-           0 0 -1 0 u3u 0 0]
-    bg2 = [d;
-           -δ2l - d;
-           κ * d;
-           -κ * d - u2l;
-           0;
-           0;
-           -u2l - κ * d;
-           u2u + κ * d]
-    bg3 = [d;
-           -δ3l - d;
-           κ * d;
-           -κ * d - u3l;
-           0;
[…]
