+++ src/LinearMPC.jl
@@ -22,7 +22,7 @@ export setup!
 export set_bounds!,add_constraint!,set_input_bounds!, set_output_bounds!
 export set_objective!, set_weights!, set_horizon!
 export set_binary_controls!
-export add_mld_constraint!, add_logic_constraint!, add_indicator_constraint!, add_product_constraint!, add_ifthenelse_constraint!
+export add_mld_constraint!, add_logic_constraint!, add_indicator_constraint!, add_product_constraint!, add_ifthenelse_constraint!, add_ifthenelse_input_constraint!
 export set_auxiliary_bounds!
 export set_disturbance!, set_x0_uncertainty!
 export set_terminal_cost!,set_prestabilizing_feedback!

[…]
