### 2026-05-23 f3632fb
Add docstrings
