+++ src/setup.jl
@@ -285,6 +285,12 @@ function require_mld(mpc::MPC)
     return mpc.mld
 end
 
+"""
+    set_auxiliary_bounds!(mpc; zmin, zmax)
+
+Sets the bounds `zmin ≤ z ≤ zmax` for the continuous auxiliary variables in an
+`MLDModel`-based controller.
+"""
 function set_auxiliary_bounds!(mpc::MPC; zmin=zeros(0), zmax=zeros(0))
     mld = require_mld(mpc)
     zmin = isempty(zmin) ? mld.zmin : float(zmin)
@@ -298,6 +304,21 @@ function set_auxiliary_bounds!(mpc::MPC; zmin=zeros(0), zmax=zeros(0))
     mpc.mpqp_issetup = false
 end
 
+"""
+    add_mld_constraint!(mpc;
+        Ax, Au, Adelta, Az, Ar, Aw, Ad, Aup, Ap,
+        Eu, Edelta, Ez, Ex, be,
+        ub, lb, ks, soft, binary, prio)
+
+Adds mixed-logical-dynamical constraints for the stage variables of an
+`MLDModel`-based controller.
+
+The constraint can either be given directly as
+`lb ≤ Ax*xₖ + Au*uₖ + Adelta*δₖ + Az*zₖ ≤ ub`, or in the standard MLD form
+`Edelta*δₖ + Ez*zₖ ≤ Ex*xₖ + Eu*uₖ + be`.
+
+`Aδ` can be used as a shorthand alias for `Adelta`.
+"""
 function add_mld_constraint!(mpc::MPC;
         Ax = nothing, Au = zeros(0,0), Adelta = zeros(0,0), Aδ = zeros(0,0), Az = zeros(0,0),
         Ar = zeros(0,0), Aw = zeros(0,0), Ad = zeros(0,0), Aup = zeros(0,0), Ap = zeros(0,0),
@@ -329,6 +350,15 @@ function add_mld_constraint!(mpc::MPC;
     return add_constraint!(mpc; Ax, Au = Ainput, Ar, Aw, Ad, Aup, Ap, ub, lb, ks, soft, binary, prio)
 end
 
+"""
+    add_logic_constraint!(mpc; Adelta, ub, lb, ks, prio)
+
+Adds inequalities involving only the binary auxiliary variables `δ`.
+
+This is a convenience wrapper around `add_mld_constraint!` for logic relations
+of the form `lb ≤ Adelta*δₖ ≤ ub`. `Aδ` can be used as a shorthand alias for
+`Adelta`.
+"""
 function add_logic_constraint!(mpc::MPC; Adelta=zeros(0,0), Aδ=zeros(0,0), ub=zeros(0), lb=zeros(0), ks = 1:mpc.Np, prio = 0)
     Adelta = isempty(Adelta) ? Aδ : Adelta
     ub = isempty(ub) ? zeros(size(Adelta,1)) : ub
@@ -336,6 +366,21 @@ function add_logic_constraint!(mpc::MPC; Adelta=zeros(0,0), Aδ=zeros(0,0), ub=z
     add_mld_constraint!(mpc; Adelta, ub, lb, ks, prio)
 end
 
+"""
+    add_indicator_constraint!(mpc, delta_id;
+        Ax, Au, c, m, M, ϵ, sense, ks, prio)
+
+Adds a big-M indicator relation between a binary variable `δ[delta_id]` and an
+affine expression in the state and control inputs.
+
+With `sense = :le`, the constraint encodes
+`δ = 1 ↔ Ax*xₖ + Au*uₖ + c ≤ 0`.
+With `sense = :ge`, it encodes
[…]
