### 2026-05-23 c324a47
Simplify
