+++ src/setup.jl
@@ -94,7 +94,7 @@ Adds the constraints lb ≤ C x  ≤ ub for the time steps k ∈ ks
 function set_output_bounds!(mpc::MPC; ymin=zeros(0), ymax=zeros(0), ks = 2:mpc.Np, soft = true, binary=false, prio = 0)
     lb = !isempty(ymin) ? ymin-mpc.model.h_offset : zeros(0)
     ub = !isempty(ymax) ? ymax-mpc.model.h_offset : zeros(0)
-    add_constraint!(mpc, Ax = mpc.model.C, Au = mpc.model.D, Ad = mpc.model.Dd, lb = lb, ub = ub; ks,soft,binary,prio)
+    add_constraint!(mpc, Ax = mpc.model.C, Ad = mpc.model.Dd, lb = lb, ub = ub; ks,soft,binary,prio)
 end
 
 """
@@ -232,6 +232,187 @@ function move_block!(mpc,blocks::Vector{<:AbstractVector{<:Number}})
     mpc.mpqp_issetup = false
 end
 
+"""
+    add_logic_constraint!(mpc; delta_ids, Adelta, ub, lb, ks, prio)
+
+Adds inequalities involving only controls that have already been marked as
+binary with `set_binary_controls!`.
+
+This is a convenience wrapper for relations of the form
+`lb ≤ Adelta*δₖ ≤ ub`. `Aδ` can be used as a shorthand alias for `Adelta`.
+The same constraints could also be added directly with `add_constraint!`.
+"""
+function add_logic_constraint!(mpc::MPC; delta_ids, Adelta=zeros(0,0), Aδ=zeros(0,0), ub=zeros(0), lb=zeros(0), ks = 1:mpc.Np, prio = 0)
+    Adelta = isempty(Adelta) ? Aδ : Adelta
+    delta_ids = _validate_input_ids(delta_ids, mpc.model.nu, "delta_ids")
+    ub = isempty(ub) ? zeros(size(Adelta,1)) : ub
+    lb = isempty(lb) ? fill(-1e30, size(Adelta,1)) : lb
+    nrows = max(length(lb), length(ub), size(Adelta, 1))
+    nrows == 0 && return
+    Au = _expand_input_block(Adelta, delta_ids, mpc.model.nu, nrows, "Adelta")
+    ub = length(ub) == nrows ? ub : [ub; fill(1e30, nrows - length(ub))]
+    lb = length(lb) == nrows ? lb : [lb; fill(-1e30, nrows - length(lb))]
+    add_constraint!(mpc; Au, ub, lb, ks, prio)
+end
+
+"""
+    add_indicator_constraint!(mpc, delta_id;
+        Ax, Au, c, m, M, ϵ, sense, ks, prio)
+
+Adds a big-M indicator relation between a control `u[delta_id]` and an affine
+expression in the state and controller inputs. The selected input should
+already be marked as binary with `set_binary_controls!`.
+
+With `sense = :le`, the constraint encodes
+`u[delta_id] = 1 ↔ Ax*xₖ + Au*uₖ + c ≤ 0`.
+With `sense = :ge`, it encodes
+`u[delta_id] = 1 ↔ Ax*xₖ + Au*uₖ + c ≥ 0`.
+
+The scalars `m` and `M` are the lower and upper big-M bounds for the affine
+expression, and `ϵ` is the strictness margin used in the reverse implication.
+"""
+function add_indicator_constraint!(mpc::MPC, delta_id::Integer;
[…]
