+++ src/mpc2mpqp.jl
@@ -143,7 +143,6 @@ function parameter_preview_direct(mpc::MPC, A, ks, Np, np)
 end
 
 stage_parameter_matrix(mpc::MPC, A, N) = mpc.settings.parameter_preview ? kron(Matrix{Float64}(I, N, N), A) : repeat(A, N, 1)
-control_stage_matrix(Np::Int, Nc::Int) = [Matrix{Float64}(I, Nc, Nc); zeros(max(0, Np + 1 - Nc), Nc)]
 
 function get_parameter_dims(mpc::MPC)
     # Use stored values if QP is set up, otherwise compute from settings
@@ -414,10 +413,17 @@ function create_objective(mpc::MPC,F,Φ,Γ,C,w::MPCWeights,nu::Int,nx::Int)
     Q_full,Qf_full = Q[1:ny,1:ny],Qf[1:ny,1:ny]
     C_full = C[1:ny,:]
 
+    pos_ids_Q = findall(diag(Q).>0); # Ignore zero indices... (and negative)
+    Q = Q[pos_ids_Q,pos_ids_Q];
+    Cp = C[pos_ids_Q,:];
+
+    pos_ids_Qf = findall(diag(Qf).>0); # Ignore zero indices... (and negative)
+    Qf = Qf[pos_ids_Qf,pos_ids_Qf];
+    Cf = C[pos_ids_Qf,:];
+
+
     # Get parameter dimensions
     nxp, nrp, ndp, nup, npp = get_parameter_dims(mpc)
-    eyeU = control_stage_matrix(N, Nc)
-    Umap = kron(eyeU[1:N, :], Matrix{Float64}(I, nu, nu))
 
     # ==== From u' R u ====
     H = kron(I(Nc),R);
@@ -437,29 +443,19 @@ function create_objective(mpc::MPC,F,Φ,Γ,C,w::MPCWeights,nu::Int,nx::Int)
         end
     end
 
-    # ==== From (Cx + Du)'Q(Cx + Du) ====
-    Dpad = [mpc.model.D; zeros(size(C,1)-ny, nu)]
-    CY = kron(I(N+1), C)
-    DY = kron(eyeU, Dpad)
-    QY = kron(I(N+1), w.Q)
-    QY[end-size(w.Qf,1)+1:end,end-size(w.Qf,2)+1:end] .= w.Qf
-    YU = CY*Γ + DY
-    Yθ = CY*Φ
-    YU_phys = kron(I(N+1), C_full) * Γ + kron(eyeU, mpc.model.D)
+    # ==== From (Cx)'Q(Cx) ====
+    CQCtot  = kron(I(N+1),Cp'*Q*Cp);
+    CQCf = Cf'*Qf*Cf
+    CQCf[1:nxp,1:nxp] .+= mpc.weights.Qfx
+    CQCtot[end-nx+1:end,end-nx+1:end].=CQCf
 
-    H += YU' * QY * YU
-    f_theta  = YU' * QY * Yθ
-    H_theta  = Yθ' * QY * Yθ
-    if !iszero(mpc.weights.Qfx)
-        XtermU = Γ[end-nx+1:end,:]
-        Xtermθ = Φ[end-nx+1:end,:]
-        H += XtermU' * mpc.weights.Qfx * XtermU
-        f_theta += XtermU' * mpc.weights.Qfx * Xtermθ
-        H_theta += Xtermθ' * mpc.weights.Qfx * Xtermθ
-    end
-    if !mpc.settings.reference_tracking && (!iszero(mpc.model.xo) || !iszero(mpc.model.uo))
[…]
