### 2026-05-23 1058f47
And support for MLD systems (#106)
* First attempt for MLD support

* Refine notation

* Remove MLD migration test

* Add docstrings

* Simplify
