+++ src/mpc_examples.jl
@@ -46,6 +46,16 @@ const _MPC_EXAMPLE_SPECS = (
         aliases = ("invpend_contact",),
         defaults = (Np = 10, Nc = 10),
     ),
+    pwa_rotation = (
+        name = "PWA Rotation System",
+        aliases = ("pwa_rotation", "rotation_mld"),
+        defaults = (Np = 3, Nc = 3),
+    ),
+    hybrid_spring = (
+        name = "Hybrid Spring-Mass-Damper",
+        aliases = ("hybrid_spring", "spring_mass_hybrid"),
+        defaults = (Np = 3, Nc = 3),
+    ),
     ballplate = (
         name = "Ball and Plate",
         aliases = ("ball", "ballplate"),
@@ -104,6 +114,41 @@ function _merge_example_kwargs(params, kwargs)
         merged[key] = value
     end
     return merged
+end
+
+function _set_example_settings!(mpc::MPC, settings, reference_tracking::Bool)
+    if isnothing(settings)
+        mpc.settings.reference_tracking = reference_tracking
+    else
+        mpc.settings = settings
+    end
+end
+
+function _affine_bounds(Ax::AbstractVector, Au::AbstractVector, c::Real,
+                        xbounds::AbstractVector, ubounds::AbstractVector)
+    span = sum(abs.(Ax) .* xbounds) + sum(abs.(Au) .* ubounds)
+    center = Float64(c)
+    return center - span, center + span
+end
+
+function _add_binary_mode_relation!(mpc::MPC, mode_id::Integer, xhigh_id::Integer, u_id::Integer;
+                                    xhigh::Bool, uhigh::Bool, ks = 1:mpc.Np)
+    A1 = zeros(1, mpc.model.nu)
+    A1[mode_id] = 1.0
+    A1[xhigh_id] = xhigh ? -1.0 : 1.0
+    add_constraint!(mpc; Au = A1, ub = [xhigh ? 0.0 : 1.0], ks)
+
+    A2 = zeros(1, mpc.model.nu)
+    A2[mode_id] = 1.0
+    A2[u_id] = uhigh ? -1.0 : 1.0
+    add_constraint!(mpc; Au = A2, ub = [uhigh ? 0.0 : 1.0], ks)
+
+    A3 = zeros(1, mpc.model.nu)
+    A3[mode_id] = 1.0
+    A3[xhigh_id] = xhigh ? -1.0 : 1.0
+    A3[u_id] = uhigh ? -1.0 : 1.0
+    lb = (xhigh ? 0.0 : 1.0) + (uhigh ? 0.0 : 1.0) - 1.0
+    add_constraint!(mpc; Au = A3, lb = [lb], ks)
 end
 
 function _build_mpc_example(::Val{:invpend}, Np, Nc; settings=nothing, kwargs...)
@@ -344,7 +389,7 @@ function _build_mpc_example(::Val{:invpend_contact}, Np, Nc; settings=nothing, k
          1 / (mc * l) -1 / (mp * l) 1 / (mp * l)]
     B = [B zeros(4, 4)]
 
-    C, D = Matrix{Float64}(I, 4, 4), zeros(4, 1)
+    C = Matrix{Float64}(I, 4, 4)
     Ts = 0.05
     F, G = zoh(A, B, Ts)
 
@@ -354,14 +399,9 @@ function _build_mpc_example(::Val{:invpend_contact}, Np, Nc; settings=nothing, k
     Rr = zeros(7)
     Qf, ~ = ared(mpc.model.F, mpc.model.G[:, 1], mpc.weights.R[1:1, 1:1], mpc.weights.Q)
[…]
