+++ test/runtests.jl
@@ -39,6 +39,10 @@ Random.seed!(1234)
         LinearMPC.mpc2mpqp(mpc)
         mpc,range = LinearMPC.mpc_examples("invpend_contact",6,6, params=Dict(:nwalls=>1));
         LinearMPC.mpc2mpqp(mpc)
+        mpc,range = LinearMPC.mpc_examples("pwa_rotation");
+        LinearMPC.mpc2mpqp(mpc)
+        mpc,range = LinearMPC.mpc_examples("hybrid_spring");
+        LinearMPC.mpc2mpqp(mpc)
         mpc,range = LinearMPC.mpc_examples("satellite");
         LinearMPC.mpc2mpqp(mpc)
         mpc,range = LinearMPC.mpc_examples("ballplate");
@@ -70,6 +74,93 @@ Random.seed!(1234)
         mpc,range = LinearMPC.mpc_examples("invpend")
         control = LinearMPC.compute_control(mpc,[5.0;5;0;0])
         @test norm(control.-1.7612519326) < 1e-6
+    end
+
+    @testset "Indicator + Product constraints" begin
+        mpc = LinearMPC.MPC([-0.8;;], [1.0 0.0 1.6]; C=[1.0;;], Np=1, Nc=1)
+        set_input_bounds!(mpc; umin=[-1.0, 0.0, -10.0], umax=[1.0, 1.0, 10.0])
+        set_binary_controls!(mpc, [2])
+        set_objective!(mpc; Q=[0.0], Qf=[0.0], R=[1e-6, 1e-6, 1e-6], eu=[0.0, 0.0, -1.0])
+        add_indicator_constraint!(mpc, 2; Ax=[-1.0;;], Au=zeros(1, 3), m=-10.0, M=10.0, ϵ=0.0)
+        add_product_constraint!(mpc, 3, 2; Ax=[1.0;;], Au=zeros(1, 3), m=-10.0, M=10.0)
+
+        x_pos = [3.0]
+        x_neg = [-2.0]
+        u_pos = compute_control(mpc, x_pos)
+        u_neg = compute_control(mpc, x_neg)
+        xplus_pos = mpc.model.F * x_pos + mpc.model.G * u_pos
+        xplus_neg = mpc.model.F * x_neg + mpc.model.G * u_neg
+        mpqp = LinearMPC.mpc2mpqp(mpc)
+
+        @test length(u_pos) == 3
+        @test mpqp.has_binaries
+        @test u_pos[3] ≈ x_pos[1] atol=1e-5
+        @test abs(u_neg[3]) < 1e-5
+        @test xplus_pos[1] ≈ 0.8 * x_pos[1] atol=1e-5
+        @test xplus_neg[1] ≈ -0.8 * x_neg[1] atol=1e-5
+    end
+
+    @testset "Indicator + ifthenelse constraints" begin
+        alpha = pi / 3
+        c, s = cos(alpha), sin(alpha)
+        function build_rotation_helper(target)
+            mpc = LinearMPC.MPC(zeros(2, 2), [0.0 1.0 0.0; 0.0 0.0 1.0];
+                                C=Matrix{Float64}(I, 2, 2), Np=1, Nc=1)
+            set_input_bounds!(mpc; umin=[0.0, -6.0, -6.0], umax=[1.0, 6.0, 6.0])
+            set_binary_controls!(mpc, [1])
+            set_objective!(mpc; Q=[0.0, 0.0], Qf=[0.0, 0.0], R=[1e-6, 1e-6, 1e-6],
+                           eu=[0.0, -target[1], -target[2]])
+            add_indicator_constraint!(mpc, 1; Ax=[1.0 0.0], Au=zeros(1, 3), m=-5.0, M=5.0, ϵ=0.0)
[…]
