+++ src/LinearMPC.jl
@@ -23,6 +23,7 @@ export set_objective!, set_weights!, set_horizon!
 export set_binary_controls!
 export set_disturbance!, set_x0_uncertainty!
 export set_terminal_cost!,set_prestabilizing_feedback!
+export add_logic_constraint!, add_indicator_constraint!, add_product_constraint!, add_ifthenelse_constraint!, add_ifthenelse_input_constraint!
 export set_state_observer!
 export set_offset_free_observer!
 export set_operating_point!, set_offset!

[…]
