### 2026-06-08 637082b
Use sentence case for common nouns
