+++ paper.md
@@ -21,11 +21,11 @@ bibliography: paper.bib
 
 # Summary
 
-`LinearMPC.jl` is a Julia package for Model Predictive Control (MPC) of linear systems. It provides a user-friendly development environment for designing, simulating, and deploying MPC controllers while targeting high-performance embedded implementations. The package bridges the gap between rapid prototyping in a high-level language and deployment on resource-constrained hardware by generating allocation-free, library-free C code that can run on any microcontroller. `LinearMPC.jl` supports both online optimization via the dual active-set Quadratic Programming (QP) solver DAQP [@Arnstrom:2022] and explicit MPC solutions computed by the multi-parametric QP solver ParametricDAQP [@Arnstrom:2024]. Additional features include robust MPC with constraint tightening, state estimation via Kalman filters, hybrid MPC with binary controls, game-theoretic MPC for Nash equilibria, and complexity certification for real-time guarantees.
+`LinearMPC.jl` is a Julia package for model predictive control (MPC) of linear systems. It provides a user-friendly development environment for designing, simulating, and deploying MPC controllers while targeting high-performance embedded implementations. The package bridges the gap between rapid prototyping in a high-level language and deployment on resource-constrained hardware by generating allocation-free, library-free C code that can run on any microcontroller. `LinearMPC.jl` supports both online optimization via the dual active-set quadratic programming (QP) solver DAQP [@Arnstrom:2022] and explicit MPC solutions computed by the multi-parametric QP solver ParametricDAQP [@Arnstrom:2024]. Additional features include robust MPC with constraint tightening, state estimation via Kalman filters, hybrid MPC with binary controls, game-theoretic MPC for Nash equilibria, and complexity certification for real-time guarantees.
 
 # Statement of need
 
-Model Predictive Control is the dominant advanced control strategy in industry, used in applications ranging from process control to autonomous vehicles and robotics. At each sampling instant, MPC solves an optimization problem that accounts for a dynamical model, constraints, and a performance objective. For linear systems, this optimization problem is a QP. Deploying MPC on embedded systems, such as microcontrollers in automotive, aerospace, and robotic applications, poses several challenges: the controller must execute within strict timing deadlines, the code must be lightweight and allocation-free, and the implementation must be verifiable.
+Model predictive control is the dominant advanced control strategy in industry, used in applications ranging from process control to autonomous vehicles and robotics. At each sampling instant, MPC solves an optimization problem that accounts for a dynamical model, constraints, and a performance objective. For linear systems, this optimization problem is a QP. Deploying MPC on embedded systems, such as microcontrollers in automotive, aerospace, and robotic applications, poses several challenges: the controller must execute within strict timing deadlines, the code must be lightweight and allocation-free, and the implementation must be verifiable.
 
 # State of the field
 
