### 2026-06-08 de1d7a1
Remove colon in heading
