### 2026-06-08 fde5cea
Make sentence parallel
