### 2026-06-10 a33bdf6
Fix various nits with bibliography, use `@online` for arXiv papers
