+++ paper.bib
@@ -80,6 +80,7 @@
     volume={140},
     pages={105676},
     year={2023},
+    doi={10.1016/j.conengprac.2023.105676},
     publisher={Elsevier}
 }
 
