### 2026-06-10 63e36c6
Fix wrong author list for acados paper
