+++ paper.bib
@@ -41,7 +41,7 @@
 }
 
 @article{Verschueren:2022,
-  author  = {Verschueren, Robin and Frison, Gianluca and Kouzoupis, Dimitris and Frey, Jonathan and van Duijkeren, Niels and Zanelli, Andrea and Nober, Brandon and Alg{\"o}wer, Thivaharan and Quirynen, Rien and Diehl, Moritz},
+  author  = {Verschueren, Robin and Frison, Gianluca and Kouzoupis, Dimitris and Frey, Jonathan and van Duijkeren, Niels and Zanelli, Andrea and Novoselnik, Branimir and Albin, Thivaharan and Quirynen, Rien and Diehl, Moritz},
   journal = {Mathematical Programming Computation},
   title   = {{{acados}}---a modular open-source framework for fast embedded optimal control},
   year    = {2022},
