+++ paper.bib
@@ -54,7 +54,7 @@
 @online{ModelPredictiveControl,
   title = {{{ModelPredictiveControl.jl}}: Advanced Process Control Made Easy in {{Julia}}},
   shorttitle = {{{ModelPredictiveControl.jl}}},
-  author = {Gagnon, Francis and Thivierge, Alex and Desbiens, André and Carlson, Fredrik Bagge},
+  author = {Gagnon, Francis and Thivierge, Alex and Desbiens, André and {Bagge Carlson}, Fredrik},
   date = {2026-05-05},
   eprint = {2411.09764},
   eprinttype = {arXiv},
