### 2026-06-10 fad40e8
Make a dedicated code generation site
