+++ docs/make.jl
@@ -155,7 +155,7 @@ makedocs(sitename = "LinearMPC.jl",
                                                    "Objective function" => "manual/objective.md",
                                                    "Constraints" =>  "manual/constraints.md",
                                                    "State observer" => "manual/observer.md",
-                                                   #"Code generation" =>  "manual/codegen.md",
+                                                   "Code generation" =>  "manual/codegen.md",
                                            ],
                               "Beyond the basics" => ["Prestabilization" => "manual/prestab.md",
                                                       "Move blocking " => "manual/moveblock.md",
