### 2026-06-13 961172d
Paper edits (#107)
* Use sentence case for common nouns

* Remove colon in heading

* Remove extra space

* Add hyphen

* Make sentence parallel

* Fix various nits with bibliography, use `@online` for arXiv papers

* Add missing DOI

* Fix wrong author list for acados paper

* Fix Fredrik Bagge Carlson author name
