+++ paper.bib
@@ -41,9 +41,9 @@
 }
 
 @article{Verschueren:2022,
-  author  = {Verschueren, Robin and Frison, Gianluca and Kouzoupis, Dimitris and Frey, Jonathan and van Duijkeren, Niels and Zanelli, Andrea and Nober, Brandon and Alg{\"o}wer, Thivaharan and Quirynen, Rien and Diehl, Moritz},
+  author  = {Verschueren, Robin and Frison, Gianluca and Kouzoupis, Dimitris and Frey, Jonathan and van Duijkeren, Niels and Zanelli, Andrea and Novoselnik, Branimir and Albin, Thivaharan and Quirynen, Rien and Diehl, Moritz},
   journal = {Mathematical Programming Computation},
-  title   = {acados---a modular open-source framework for fast embedded optimal control},
+  title   = {{{acados}}---a modular open-source framework for fast embedded optimal control},
   year    = {2022},
   volume  = {14},
   number  = {1},
@@ -51,11 +51,16 @@
   doi     = {10.1007/s12532-021-00208-8}
 }
 
-@article{ModelPredictiveControl,
-    title={{ModelPredictiveControl}. jl: advanced process control made easy in Julia},
-    author={Gagnon, Francis and Thivierge, Alex and Desbiens, Andr{\'e} and Carlson, Fredrik Bagge},
-    journal={arXiv preprint arXiv:2411.09764},
-    year={2024}
+@online{ModelPredictiveControl,
+  title = {{{ModelPredictiveControl.jl}}: Advanced Process Control Made Easy in {{Julia}}},
+  shorttitle = {{{ModelPredictiveControl.jl}}},
+  author = {Gagnon, Francis and Thivierge, Alex and Desbiens, André and {Bagge Carlson}, Fredrik},
+  date = {2026-05-05},
+  eprint = {2411.09764},
+  eprinttype = {arXiv},
+  eprintclass = {eess.SY},
+  doi = {10.48550/arXiv.2411.09764},
+  pubstate = {prepublished}
 }
 
 @ARTICLE{Arnstrom:2022b,
@@ -69,39 +74,56 @@
     doi={10.1109/TAC.2021.3090749}}
 
 @article{Fiedler:2023,
-    title={do-mpc: Towards FAIR nonlinear and robust model predictive control},
+    title={{{do-mpc}}: Towards {FAIR} nonlinear and robust model predictive control},
     author={Fiedler, Felix and Karg, Benjamin and L{\"u}ken, Lukas and Brandner, Dean and Heinlein, Moritz and Brabender, Felix and Lucia, Sergio},
     journal={Control Engineering Practice},
     volume={140},
     pages={105676},
     year={2023},
+    doi={10.1016/j.conengprac.2023.105676},
     publisher={Elsevier}
 }
 
-@article{Wikner:2026,
-    title={On Real-Time Feasibility of High-Rate MPC using an Active-Set Method on Nano-Quadcopters},
-    author={Wikner, Joel and Arnstr{\"o}m, Daniel and Axehill, Daniel},
-    journal={arXiv preprint arXiv:2603.09342},
-    year={2026}
+@online{Wikner:2026,
+  title = {On Real-Time Feasibility of High-Rate {{MPC}} Using an Active-Set Method on Nano-Quadcopters},
+  author = {Wikner, Joel and Arnström, Daniel and Axehill, Daniel},
+  date = {2026-03-10},
+  eprint = {2603.09342},
+  eprinttype = {arXiv},
+  eprintclass = {math.OC},
+  doi = {10.48550/arXiv.2603.09342},
+  pubstate = {prepublished}
 }
 
-@article{arnstrom2025prioritized,
[…]
