+++ v0.9.0-joss/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"category":"section","location":"manual/reference_preview/#Reference-Preview","page":"Reference Preview","text":"Reference preview allows the MPC controller to use knowledge of future reference values, enabling better tracking performance for time-varying setpoints.","title":"Reference Preview"},{"category":"section","location":"manual/reference_preview/#Setup","page":"Reference Preview","text":"Enable reference preview in the MPC settings:\n\nmpc.settings.reference_preview = true\nsetup!(mpc)  # Rebuild the controller with new settings","title":"Setup"},{"category":"section","location":"manual/reference_preview/#Usage","page":"Reference Preview","text":"","title":"Usage"},{"category":"section","location":"manual/reference_preview/#Reference-Format","page":"Reference Preview","text":"Provide references as a matrix of size (ny, Np) where:\n\nny is the number of outputs\nNp is the prediction horizon\n\n# For a system with 2 outputs and prediction horizon of 5\nr_trajectory = [1.0 1.5 2.0 2.0 2.0;   # Reference for output 1\n                0.0 0.0 0.5 1.0 1.0]   # Reference for output 2\n\nu = compute_control(mpc, x; r=r_trajectory)","title":"Reference Format"},{"category":"section","location":"manual/reference_preview/#Single-Reference","page":"Reference Preview","text":"Vector inputs are automatically broadcast across the prediction horizon:\n\n# This reference will be used for all time steps\nu = compute_control(mpc, x; r=[1.0, 0.0])","title":"Single Reference"},{"category":"section","location":"manual/reference_preview/#Example","page":"Reference Preview","text":"using LinearMPC\n\n# Create MPC with reference preview\nA = [1 1; 0 1]\nB = [0; 1] \nC = [1.0 0; 0 1.0]\nmpc = LinearMPC.MPC(A, B; C, Np=10, Nc=5)\nset_objective!(mpc; Q=[1.0, 1.0], R=[0.1])\n\n# Enable reference preview\nmpc.settings.reference_preview = true\nsetup!(mpc)\n\n# Define reference trajectory (2 outputs × 10 time steps)\nr_traj = [zeros(1,5) ones(1,5);      # Step change for output 1\n          zeros(1,10)]               # Zero reference for output 2\n\n# Compute control with preview\nx = randn(2)\nu = compute_control(mpc, x; r=r_traj)\nnothing # hide","title":"Example"},{"category":"section","location":"manual/reference_preview/#Simulation","page":"Reference Preview","text":"The Simulation function automatically handles reference preview when enabled:\n\n# Reference trajectory for simulation\nN_sim = 30\nr_sim = zeros(2, N_sim)\nr_sim[1, 15:end] .= 1.0  # Step reference at time 20\n\nsim = Simulation(mpc; x0=zeros(2), N=N_sim, r=r_sim)\nnothing # hide\n\nusing Plots\nplot(sim)\n\nNotice how the controller initiates the move from 0 to 1 already before the reference actually changes. If you look really closely, you can see that the output actually initially goes down rather than up before the reference change occurs, this is a phenomenon arising due to the quadratic cost function which makes it worthwhile to allow a small (\"small squared\" is really small) deviation in order to build up momentum in advance of the large reference change in order to reduce the large error (\"large squared\" is really large) arising from the step muc
[…]
