+++ v0.9.0-joss/siteinfo.js
@@ -1,0 +1,2 @@
+var DOCUMENTER_CURRENT_VERSION = "v0.9.0-joss";
+var DOCUMENTER_IS_DEV_VERSION = false;
